### 2016-06-28 #33 @whedon
Submission: RSMTool: a Python Package for facilitating research on automated scoring models
**Submitting author:** @desilinguist (<a href="http://orcid.org/0000-0001-9354-6851">Nitin Madnani</a>)
**Repository:** <a href="https://github.com/EducationalTestingService/rsmtool" target ="_blank">https://github.com/EducationalTestingService/rsmtool</a>
**Version:** v5.1.0
