### 2016-06-28 92e380c
Fixing bibtex entries
