+++ paper.bib
@@ -16,7 +16,7 @@ year = {2012}
   urldate = {2016-06-23}
 }
 
-@misc{dan_blanchard_2014_12825,
+@misc{skll,
   author       = {Dan Blanchard and
                   Nitin Madnani and
                   Michael Heilman and
@@ -32,7 +32,7 @@ year = {2012}
   url          = {http://dx.doi.org/10.5281/zenodo.12825}
 }
 
-@inproceedings{Loukina2015,
+@inproceedings{loukina2015,
 address = {Denver, Colorado},
 author = {Loukina, Anastassia and Zechner, Klaus and Chen, Lei and Heilman, Michael},
 booktitle = {Proceedings of the Tenth Workshop on Innovative Use of NLP for Building Educational Applications},
