### 2016-06-29 #33 @desilinguist
Thanks, Jeremy!
On Wed, Jun 29, 2016 at 6:29 PM Jeremy G. Kahn <notifications@github.com>
wrote:

> maybe a tentative ✋ I am reviewing this? If nobody else speaks up this
> week, I will take it on to do a review next week (July 5-8).
>
> Full disclosure: Nitin (the submitting author) and I are friendly (and see
> each other, mostly at conferences, about every 18 months).
>
> I don't know anything about this *specific* work of his.
>
