### 2016-07-06 5fd01ac
Fix License to actually declare copyright
Discovered missing when comparing to output of the `lice` tool.

part of JOSS review:
https://github.com/openjournals/joss-reviews/issues/33
