### 2016-07-07 136133f
Add DOI to Williamson paper
Per https://github.com/openjournals/joss-reviews/issues/33.
