### 2016-07-07 #33 @jkahn
General checks and references
---------------------
All the outer formatting looks good.

Might be nice to find a DOI for Loukina 2015 but I can't figure out how to get DOIs from ACLweb.
