### 2016-07-07 #33 @jkahn
Statement of need
-------------------
I'm not 100% clear what the statement of actual need is here.  I get that this is a useful tool within the ETS, but I'm not certain about what the contribution of *this* package is, and who the researchers (outside of the ETS itself) would be that would find this to be a useful tool in hand.

I'm *not* saying that those researchers don't exist -- I might even be one of them -- but I don't think the statement of need clearly reflects what a imaginary researcher Professor X would use these tools for. Perhaps a tutorial walking through Professor X's thought process would help clarify what problem this solves -- and if the tutorial was part of the documentation, so much the better.

I suspect is meeting >1 need and some researchers may have only a subset of those needs, and that is okay, but those are not clear.  (See upcoming comment about multiply-situated work.)
