### 2016-07-07 #33 @desilinguist
Please see the "available documentation" section in the main README. All
the config files for all four tools format are fully documented.

There's only example, that's true. That's because we expect the "rsmtool"
endpoint to be the most commonly used. We can certainly make that clearer.
On Thu, Jul 7, 2016 at 7:12 PM Jeremy G. Kahn <notifications@github.com>
wrote:

> Undocumented entry points confuse new users
>
> There are at least three different command-line tools (endpoints, in the
> Python jargon) that all seem to take the same argument structure (a config
> file) but presumably have different formats expected in the config files.
> There's exactly one example use, scraped from a Kaggle competition, but it
> only uses one of the CLI endpoints (rsmtool); the others (e.g. rsmeval
> and rsmcompare) don't have sample usages.
>
> This doesn't help your statement of need much, either.
>
