### 2016-07-07 #33 @jkahn
Separating the `pandas.DataFrame` manipulator APIs from the input and output formats
------------------------------------
As far as I can tell, the format of the input features is also undocumented. I don't have a clear picture of how I might (as an external developer) go about creating the datasets of "features" (a hugely overloaded term, further overloaded in the limited documentation provided here) applied to these competitions.  Furthermore, I don't understand from the documentation what aspects of those files are being displayed in the resulting HTML generation.

I wonder if a clear difference between dataframe format and on-disk storage format would clarify things here.  (I can imagine:

- a "stored features" form in JSON, documented
- an API specification about the shapes and labels of DataFrame objects for input
- some utilities to lift the stored results to appropriate API inputs
- an API specification about the shapes and labels generated for output
- some utilities (probably notebooks) to combine and display bundled API outputs

I'd thus like to see a separation among the following concerns, within the documentation, all of which seem to be at play here:

- **loading to dataframe** tools and required formats for loading input data
- **dataframe manipulators** comparing human evaluation and machine-extracted "features" to each other, including manipulations to extract useful summary statistics and dataframes
- **display utilities** convenience collections (notebooks?) to gather and display the statistics against each other
- **commandline tools** for coordinating the execution of the pieces above
