### 2016-07-08 #33 @jkahn
Recommendation: accept - Minor revisions.
I think it's clear that there's a docs problem here:
- statement of need is too high-level and making it lower-level will probably articulate >1 role of users
- doc'n is unclear about documenting external storage format vs internal APIs
- API doc'n is not easily browseable (e.g. through a sphinx `.. autodoc`-including `api.rst` file)
- Documentation not easily browseable [though github itself renders the .md files, the API documentation isn't so easy. this could be remedied quickly with http://readthedocs.io or similar]

Separately, I have a few further quibbles that should not block publication but should probably be
- I'd like to see an introductory tutorial notebook, but I don't think that should block publication of this version
- A package with this many `.py` files should probably have a `coverage` and a linter-style test run at least between version revisions.
- installing through pip is the standard for Python -- and everything here works (if a custom `requirements.txt` file is included).  It'd be nice to have `sdist` tarballs available on PyPI -- this would even be compatible with conda installations, after all, and the code defined in this package does not include any non-Python code AFAICT.  (many of the declared dependencies require careful work with compilation, but that itself should be outside the scope of this project.)
