### 2016-07-08 #33 @desilinguist
Indeed. Thanks for the careful review, Jeremy!
On Fri, Jul 8, 2016 at 13:33 Anastassia Loukina <notifications@github.com>
wrote:

> Thank you for such a detailed review, Jeremy! We'll go over your
> suggestions with Nitin.
>
