### 2016-07-18 209b3cb
Integrate codecov.io into CI tests.
