+++ circle.yml
@@ -14,6 +14,6 @@ dependencies:
 # Run test
 test:
   override:
-    - ${HOME}/miniconda3/bin/nosetests -v tests
+    - ${HOME}/miniconda3/bin/nosetests -v tests --with-coverage --cover-package=rsmtool
   post:
     - bash <(curl -s https://codecov.io/bash)
