+++ circle.yml
@@ -16,4 +16,4 @@ test:
   override:
     - ${HOME}/miniconda3/bin/nosetests -v tests --with-coverage --cover-package=rsmtool
   post:
-    - bash <(curl -s https://codecov.io/bash)
+    - pip install codecov && codecov
