### 2016-07-18 d8fe7d0
Switching to coveralls.io
- codecov.io doesn’t seem to be working correctly.
