+++ circle.yml
@@ -9,11 +9,12 @@ dependencies:
     - ${HOME}/miniconda3/bin/conda config --add channels desilinguist
     - ${HOME}/miniconda3/bin/conda update --yes conda
     - ${HOME}/miniconda3/bin/conda install --file conda_requirements.txt --yes
+    - ${HOME}/miniconda3/bin/pip install nose-cov python-coveralls
     - ${HOME}/miniconda3/bin/pip install -e .
 
 # Run test
 test:
   override:
-    - ${HOME}/miniconda3/bin/nosetests -v tests --with-coverage --cover-package=rsmtool
+    - ${HOME}/miniconda3/bin/nosetests -v tests --with-coverage --cover-package=rsmtool --cov-config .coveragerc
   post:
-    - pip install codecov && codecov
+    - coveralls --config_file .coveragerc
