### 2016-07-18 2942eee
More coveralls fiddling.
