### 2016-07-18 3a84a8e
Add coveralls badge to README
