### 2016-07-18 37cdea3
Various edits to readme and other docs
