### 2016-07-18 5f7cac6
minor typos
