### 2016-07-19 58fe2c4
changes to readme
