### 2016-07-19 085a959
Changes to the notebooks to allow for missing data files
