### 2016-07-19 0bbaf87
Examples for all three tools
