### 2016-07-19 0bf2bb8
Minor edits to readme
