+++ README.md
@@ -19,10 +19,10 @@ Note that although the analyses included into the tool as well as wording of the
 RSMTool has two main use cases illustrated by the following examples:
 
 * Use case 1: a researcher *has* a set of responses such as essays or recorded spoken responses which have already been assigned numeric scores. She also processed these responses and extracted a set of numeric features either using systems such as Coh-Metrix, TextEvaluator, OpenSmile or by manually creating a new set of features. 
-She uses *RSMTool* to build a regression-based system which predicts the scores based on the extracted features using different types of regressions, to compute descriptive statistics for all features, to generate scores for a held-out evaluation set, and to obtain a comprehensive evaluation of system performance. While she could use `R`, `SPSS` or other tools to perform these analyses, `RSMTool` allows her to do so by running a single command. Furthermore, the analyses included into the tool highlight the issues that are commonly considered w need to be considered when building automated scoring models. She can then use the tool outputs for further exploratory analysis of her model. 
+She uses *RSMTool* to build a regression-based system which predicts the scores based on the extracted features using different types of regressions, to compute descriptive statistics for all features, to generate scores for a held-out evaluation set, and to obtain a comprehensive evaluation of system performance. While she could use `R`, `SPSS` or other tools to perform these analyses, `RSMTool` allows her to do so by running a single command. Furthermore, the analyses included into the tool highlight the issues that need to be considered when building automated scoring models. She can then use the tool outputs for further exploratory analysis of her model. 
 
 * Use case 2: A researcher *has* an automated scoring system for grading short responses that extracts the features and computes the score. He wants to evaluate the system performance using metrics commonly used in educational community but not always available in standard machine learning packages as well as to conduct additional analyses to evaluate system fairness and compare system performance to human-human agreement.
-He *uses* RSMTool to set up customized evaluation report using a combination of existing and custom sections and then runs RSMEval to quickly produce a new report for each version of his system. 
+He *uses* RSMTool to set up customized evaluation report using a combination of existing and custom sections and then runs RSMEval to quickly produce a new report for each version of his system.
 
 
 RSMTool provides the following main scripts. The documentation for each of these tools is provided separately (see below). 
