### 2016-07-19 e020db6
fixed the test
