### 2016-07-20 9c356bb
Start working on readthedocs documentation
