+++ doc/usage_rsmtool.rst
@@ -1,0 +1,178 @@
+Using RSMTool
+=============
+
+For most users, the primary means of using RSMTool will be via the command-line utility ``rsmtool``. We refer to each run of ``rsmtool`` as an "experiment".
+
+Input
+-----
+
+``rsmtool`` requires the following two inputs to run an experiment:
+
+1. Data files for training and evaluation sets in ``.csv`` format. Each row should correspond to a single response and contain numeric feature values extracted for this response. In addition, there should be a column with a unique identifier (ID) for each response and a column with the human score for each response. The names for these two columns are defined in the configuration file.
+
+2. A configuration file in ``.json`` format with paths to the training and evaluation files, the type of regressor to use, and other information required by ``rsmtool``. See :ref:`experiment configuration file <config_file>` for more details.
+
+By default, ``rsmtool`` will use all of the features present in the training CSV file. If you want to use a specific set of features, you need to provide a second ``.json`` file specifying the list of features. See :ref:`feature file <feature_file>` for more details.
+
+``rsmtool`` does the following:
+
+preprocesses the training data (outlier truncation, normalization), learns a scoring model on the pre-processed training data, and generates a report with the description of the model and in-depth evaluation of its performance on the evaluation data.
+
+
+.. _config_file:
+
+Experiment configuration file
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+This file provides an overall configuration for the experiment. The file must be in ``.json`` format. Here's an example of a configuration file.
+
+<example>
+
+There are four required fields.
+
+experiment_id
+"""""""""""""
+An identifier for the experiment that will be used to name the report and all intermediate CSV files. It can be any combination of alphanumeric values and must *not* contain spaces.
+
+model
+"""""
+The machine learner you want to use to build the regressi. See [available models](available_models.md) for the list of available learners.
+
+train_file
+""""""""""
+The path to the training data in ``.csv`` format. Can be absolute or relative to the location of config file.
+
+test_file
+"""""""""
+The path to the evaluation data in ``.csv`` format. Can be absolute or relative to the location of config file.
+
+description *(Optional)*
+""""""""""""""""""""""""
+A brief description of the experiment. This will be included in the report. The description can contain spaces and punctuation. This is blank by default.
+
+id_column *(Optional)*
+""""""""""""""""""""""
+The name of the column containing the response IDs. Defaults to ``spkitemid``, i.e., if this is not specified, ``rsmtool`` will look for a column called ``spkitemid`` in the training and evaluation files.
+
+.. _train_label_column:
+
+train_label_column *(Optional)*
+"""""""""""""""""""""""""""""""
+The name for the column containing the human scores in the training data. If set to to ``fake``, fake scores will be generated using randomly sampled integers. This option may be useful if you only need descriptive statistics for the data and do not care about the other analyses. Defaults to ``sc1``.
+
+.. _test_label_column:
+
+test_label_column *(Optional)*
+""""""""""""""""""""""""""""""
+The name for the column containing the human scores in the training data. If set to to ``fake``, fake scores will be generated using randomly sampled integers. This option may be useful if you only need descriptive statistics for the data and do not care about the other analyses. Defaults to ``sc1``.
+
+.. note::
+
+    All responses with non-numeric values in either ``train_label_column`` or ``test_label_column`` and/or those with non-numeric values for relevant features will be automatically excluded from model training and evaluation.
+
+
+features *(Optional)*
+"""""""""""""""""""""
+The path to the ``.json`` file containing the list of features and transformations, if any . Can be absolute or relative to the location of config file.
+
+.. note::
+
[…]
