### 2016-07-20 4ee3981
Fix duplicated list in `model.py`.
