+++ doc/usage_rsmtool.rst
@@ -8,16 +8,9 @@ Input
 
 ``rsmtool`` requires the following two inputs to run an experiment:
 
-1. Data files for training and evaluation sets in ``.csv`` format. Each row should correspond to a single response and contain numeric feature values extracted for this response. In addition, there should be a column with a unique identifier (ID) for each response and a column with the human score for each response. The names for these two columns are defined in the configuration file.
+1. Data files for training and evaluation sets in ``.csv`` format. Each row should correspond to a single response and contain numeric feature values extracted for this response. In addition, there should be a column with a unique identifier (ID) for each response and a column with the human score for each response.
 
 2. A configuration file in ``.json`` format with paths to the training and evaluation files, the type of regressor to use, and other information required by ``rsmtool``. See :ref:`experiment configuration file <config_file>` for more details.
-
-By default, ``rsmtool`` will use all of the features present in the training CSV file. If you want to use a specific set of features, you need to provide a second ``.json`` file specifying the list of features. See :ref:`feature file <feature_file>` for more details.
-
-``rsmtool`` does the following:
-
-preprocesses the training data (outlier truncation, normalization), learns a scoring model on the pre-processed training data, and generates a report with the description of the model and in-depth evaluation of its performance on the evaluation data.
-
 
 .. _config_file:
 
@@ -28,7 +21,7 @@ This file provides an overall configuration for the experiment. The file must be
 
 <example>
 
-There are four required fields.
+There are four required fields and rest are all optional.
 
 experiment_id
 """""""""""""
@@ -36,7 +29,7 @@ An identifier for the experiment that will be used to name the report and all in
 
 model
 """""
-The machine learner you want to use to build the regressi. See [available models](available_models.md) for the list of available learners.
+The machine learner you want to use to build the scoring model. Possible values include :doc:`built-in linear regression models <builtin_models>` as well as all of the regressors available via `SKLL <http://skll.readthedocs.io/en/latest/run_experiment.html#learners>`_.
 
 train_file
 """"""""""
@@ -49,6 +42,14 @@ The path to the evaluation data in ``.csv`` format. Can be absolute or relative 
 description *(Optional)*
 """"""""""""""""""""""""
 A brief description of the experiment. This will be included in the report. The description can contain spaces and punctuation. This is blank by default.
+
+features *(Optional)*
+"""""""""""""""""""""
+By default, ``rsmtool`` will use all of the features present in the training and evaluation CSV files. If you want to use a specific set of features, you need to provide a second ``.json`` file specifying the list of features. The value for this field is the path to this ``.json`` file. It can be absolute or relative to the location of config file.
+
+.. note::
+
+    See :doc:`selecting features <feature_selection>` for more details.
 
 id_column *(Optional)*
 """"""""""""""""""""""
@@ -70,15 +71,7 @@ The name for the column containing the human scores in the training data. If set
 
     All responses with non-numeric values in either ``train_label_column`` or ``test_label_column`` and/or those with non-numeric values for relevant features will be automatically excluded from model training and evaluation.
 
-
-features *(Optional)*
-"""""""""""""""""""""
-The path to the ``.json`` file containing the list of features and transformations, if any . Can be absolute or relative to the location of config file.
-
-.. note::
-
-    By default, ``rsmtool`` uses *all* columns in the training file as features except for the ones with the following names: ``spkitemid``, ``spkitemlab``, ``itemType``, `r1`, `r2`, `score`, `sc`, `sc1`, `adj`, and the column names specified in the config file  (e.g. `length_column`, `subgroups` as well as `train_label_column` and `test_label_column`). The final set of features will be saved in the features/ folder.
-
+.. _length_column:
 
 length_column *(Optional)*
[…]
