+++ doc/intermediate_files_rsmtool.rst
@@ -1,0 +1,206 @@
+.. _intermediate_files_rsmtool:
+
+Intermediate CSV files
+----------------------
+
+Although the primary output of RSMTool is an HTML report, we also want the user to be able to conduct additional analyses outside of RSMTool. To this end, all of the tables produced in an RSMTool experiment report are saved as ``.csv`` files in the ``output`` directory.
+
+The names of all files begin with the ``experiment_id`` provided by the user in the experiment configuration file. In addition, the names for certain columns are set to default values in these files irrespective of what they were named in the original ``.csv`` files. This is because RSMTool standardizes these column names internally for convenience. These values are:
+
+- `spkitemid` for the column containing response IDs.
+- `sc1` for the column containing the human scores used as training labels.
+- `sc2` for the column containing the second human scores, if this column was specified in the configuration file.
+- `length` for the column containing response length, if this column was specified in the configuration file.
+- `candidate` for the column containing candidate IDs, if this column was specified in the configuration file.
+
+
+Features values
+^^^^^^^^^^^^^^^
+filenames: ``train_features.csv``, ``test_features.csv``, ``train_preprocessed_features.csv``, ``test_preprocessed_features.csv``
+
+These files contain the raw and pre-processed feature values for the training and evaluation sets. They include *only* includes the rows that were used for training/evaluating the models after filtering. For models with feature selection, these files *only* include the features that ended up being included in the model.
+
+.. note::
+
+    By default RSMTool filters out non-numeric feature values and non-numeric/zero human scores from both the training and evaluation sets. Zero scores can be kept by setting the `exclude_zero_scores` to `false`.
+
+Flagged responses
+^^^^^^^^^^^^^^^^^
+filenames: ``train_responses_with_excluded_flags.csv``, ``test_responses_with_excluded_flags.csv``
+
+These files contain all of the rows in the training and evaluation sets that were filtered out based on conditions specified in :ref:`flag_column <flag_column>`.  Note that the training/evaluation ``.csv`` files contained columns with internal names such as ``sc1`` or ``length`` but these columns were not actually used by ``rsmtool``, these columns will also be included into these files but their names will be changed to ``##name##`` (e.g. ``##sc1##``).
+
+Excluded responses
+^^^^^^^^^^^^^^^^^^
+filenames: ``train_excluded_responses.csv``, ``test_excluded_responses.csv``
+
+These files contain all of the rows in the training and evaluation sets that were filtered out because of feature values or scores. For models with feature selection, these files *only* include the features that ended up being included in the model.
+
+Response metadata
+^^^^^^^^^^^^^^^^^
+filenames: ``train_metadata.csv``, ``test_metadata.csv``
+
+These files contain the metadata columns (``id_column``,  ``subgroups`` if provided) for the rows in the training and evaluation sets that were *not* excluded for some reason.
+
+Unused columns
+^^^^^^^^^^^^^^
+filenames: ``train_other_columns.csv``, ``test_other_columns.csv``
+
+These files contain all of the the columns from the original features files that are not present in the ``*_feature.csv`` and ``*_metadata.csv`` files. They only includes the rows from the training and evaluation sets that were not filtered out. Note that the training/evaluation ``.csv`` files contained columns with internal names such as ``sc1`` or ``length`` but these columns were not actually used by ``rsmtool``, these columns will also be included into these files but their names will be changed to ``##name##`` (e.g. ``##sc1##``).
+
+Response length
+^^^^^^^^^^^^^^^
+filename: ``train_response_lengths.csv``
+
[…]
