### 2016-07-21 a01226a
Prevent duplicate warnings while building.
