+++ doc/conf.py
@@ -80,7 +80,10 @@ release = version
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This patterns also effect to html_static_path and html_extra_path
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
+exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store',
+                    'config_rsmtool.rst', 'builtin_models.rst',
+                    'feature_selection_rsmtool.rst',
+                    'intermediate_files_rsmtool.rst']
 
 # The reST default role (used for this markup: `text`) to use for all
 # documents.
