+++ doc/usage_rsmtool.rst
@@ -8,189 +8,7 @@ Input
 
 ``rsmtool`` requires a single input to run an experiment: a configuration file. It can also take an output directory as an optional second argument. If the latter is not specified, ``rsmtool`` will use the current directory as the output directory.
 
-
-.. _config_file:
-
-Experiment configuration file
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-This is a file in ``.json`` format that provides overall configuration options for an ``rsmtool`` experiment. An example configuration file can be found here.
-
-There are four required fields and rest are all optional.
-
-experiment_id
-"""""""""""""
-An identifier for the experiment that will be used to name the report and all intermediate CSV files. It can be any combination of alphanumeric values and must *not* contain spaces.
-
-model
-"""""
-The machine learner you want to use to build the scoring model. Possible values include :ref:`built-in linear regression models <builtin_models>` as well as all of the regressors available via `SKLL <http://skll.readthedocs.io/en/latest/run_experiment.html#learners>`_.
-
-train_file
-""""""""""
-The path to the training data feature file in ``.csv`` format. Each row should correspond to a single response and contain numeric feature values extracted for this response. In addition, there should be a column with a unique identifier (ID) for each response and a column with the human score for each response. Can be absolute or relative to the location of config file.
-
-test_file
-"""""""""
-The path to the evaluation data feature file in ``.csv`` format. Each row should correspond to a single response and contain numeric feature values extracted for this response. In addition, there should be a column with a unique identifier (ID) for each response and a column with the human score for each response. Can be absolute or relative to the location of config file.
-
-description *(Optional)*
-""""""""""""""""""""""""
-A brief description of the experiment. This will be included in the report. The description can contain spaces and punctuation. This is blank by default.
-
-features *(Optional)*
-"""""""""""""""""""""
-By default, ``rsmtool`` will use all of the features present in the training and evaluation CSV files. If you want to use a specific set of features, you need to provide a second ``.json`` file specifying the list of features. The value for this field is the path to this ``.json`` file. It can be absolute or relative to the location of config file.
-
-.. note::
-
-    See :doc:`selecting features <feature_selection>` for more details.
-
-id_column *(Optional)*
-""""""""""""""""""""""
-The name of the column containing the response IDs. Defaults to ``spkitemid``, i.e., if this is not specified, ``rsmtool`` will look for a column called ``spkitemid`` in the training and evaluation files.
-
-.. _train_label_column:
-
-train_label_column *(Optional)*
-"""""""""""""""""""""""""""""""
-The name for the column containing the human scores in the training data. If set to to ``fake``, fake scores will be generated using randomly sampled integers. This option may be useful if you only need descriptive statistics for the data and do not care about the other analyses. Defaults to ``sc1``.
-
-.. _test_label_column:
-
-test_label_column *(Optional)*
-""""""""""""""""""""""""""""""
-The name for the column containing the human scores in the training data. If set to to ``fake``, fake scores will be generated using randomly sampled integers. This option may be useful if you only need descriptive statistics for the data and do not care about the other analyses. Defaults to ``sc1``.
-
-.. note::
-
-    All responses with non-numeric values in either ``train_label_column`` or ``test_label_column`` and/or those with non-numeric values for relevant features will be automatically excluded from model training and evaluation.
-
-.. _length_column:
-
-length_column *(Optional)*
-""""""""""""""""""""""""""
[…]
