+++ doc/feature_selection_rsmtool.rst
@@ -1,0 +1,4 @@
+.. _feature_selection:
+
+Selecting Features
+------------------
