### 2016-07-21 14197c4
Various minor edits
