### 2016-07-21 c1e41fa
[ci skip] More changes to docs
- including changes from @aloukina.
