+++ doc/feature_selection_rsmtool.rst
@@ -2,3 +2,5 @@
 
 Selecting Features
 ------------------
+
+By default, ``rsmtool`` will use all of the features provided in the training and evaluation ``.csv`` files. However, it is possible for the user to either manually choose a subset of features and
