### 2016-07-21 d0af84c
[ci skip] add skeleton for advanced usage.
