+++ doc/advanced_usage.rst
@@ -1,0 +1,10 @@
+Advanced Uses of RSMTool
+------------------------
+
+In addition to training and evaluating regression-based scoring models, RSMTool provides three other command-line scripts for more advanced users.
+
+.. include:: usage_rsmeval.rst
+
+.. include:: usage_rsmpredict.rst
+
+.. include:: usage_rsmcompare.rst
