### 2016-07-21 b01cbf2
documentation for rsmeval [skip ci]
