+++ doc/config_rsmeval.rst
@@ -1,0 +1,151 @@
+Experiment configuration file
+"""""""""""""""""""""""""""""
+
+This is a file in ``.json`` format that provides overall configuration options for an ``rsmeval`` experiment. An example configuration file can be found here.
+
+There are four required fields and the rest are all optional.
+
+experiment_id
+~~~~~~~~~~~~~
+An identifier for the experiment that will be used to name the report and all intermediate CSV files. It can be any combination of alphanumeric values and must *not* contain spaces.
+
+predictions_file
+~~~~~~~~~~~~~~~~
+The path to the file with predictions in ``.csv`` format. Can be absolute or relative to the location of config file.  Each row should correspond to a single response and contain the predicted and observed scores for this response. In addition, there should be a column with a unique identifier (ID) for each response. The path can be absolute or relative to the location of config file.
+
+system_score_column
+~~~~~~~~~~~~~~~~~~~
+The name for the column containing the system scores (predicted scores). These scores will be used for evaluation. 
+
+trim_min
+~~~~~~~~
+The single numeric value for the lowest possible score that the machine should predict. This value will be used to compute trimmed (bound) machine scores.
+
+trim_max
+~~~~~~~~
+
+The single numeric value for the highest possible score that the machine should predict. This value will be used to compute trimmed (bound) machine scores. 
+
+.. note::
+
+    ``trim_min`` and ``trim_max`` fields are optional fields for ``rsmtool`` but are required fields for ``rsmeval``.
+
+description *(Optional)*
+~~~~~~~~~~~~~~~~~~~~~~~~
+
+A brief description of the experiment. This will be included in the report. The description can contain spaces and punctuation. This is blank by default.
+
+
+
+id_column *(Optional)*
+~~~~~~~~~~~~~~~~~~~~~~
+
+The name of the column containing the response IDs. Defaults to ``spkitemid``, i.e., if this is not specified, ``rsmeval`` will look for a column called ``spkitemid`` in the prediction file.
+
+
+human_score_column *(Optional)*
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+The name for the column containing the human scores for each response. The values in this column will be used as observed scores. Defaults to ``sc1``. 
+
+
+.. note::
+
+    All responses with non-numeric values or zeros in either ``human_score_column`` or ``system_score_column`` will be automatically excluded from model training and evaluation. You can use :ref:`exclude_zero_scores_eval` to keep responses with zero scores. 
+
+
+second_human_score_column *(Optional)*
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+The name for an optional column in the test data containing a second human score for each response. If specified, additional information about human-human agreement and degradation will be computed and included in the report. Note that this column must contain either numbers or be empty. Non-numeric values are *not* accepted. Note also that the :ref:`exclude_zero_scores_eval` option below will apply to this column too.
+
+.. _flag_column_eval:
+
+flag_column *(Optional)*
+~~~~~~~~~~~~~~~~~~~~~~~~
+
+This field makes it possible to only use responses with particular values in a given column (e.g. only responses with a value of ``0`` in a column called ``ADVISORY``). The field takes a dictionary in Python format where the keys are the names of the columns and the values are lists of values for responses that will be used to train the model. For example, a value of ``{"ADVISORY": 0}`` will mean that ``rsmtool`` will *only* use responses for which the ``ADVISORY`` column has the value 0. Defaults to ``None``.
+
+.. note::
+
[…]
