### 2016-07-21 01b536b
doc for rsmpredict
