+++ doc/feature_selection_rsmtool.rst
@@ -5,6 +5,12 @@ Selecting Features
 
 By default, ``rsmtool`` will use all of the features provided in the training and evaluation ``.csv`` files. However, it is possible for you to either manually choose a subset of features or to have ``rsmtool`` perform automatic feature selection.
 
+.. note::
+
+    For all feature selection methods, the final set of features will be saved in the ``feature`` folder in the experiment output directory.
+
+.. _manual_feature_selection:
+
 Manual Feature selection
 ^^^^^^^^^^^^^^^^^^^^^^^^
 To manually select a subset of features, you must provide a ``.json`` file specifying information about what features should be included in the final scoring model. For additional flexibility, the same file also allows you to describe transformations to be applied to the raw features before being included in the model.
@@ -34,22 +40,20 @@ There are three required fields.
 
 feature
 """""""
-The name of the feature. This must match the feature name in the training and evaluation ``.csv`` files exactly, including capitalization. Feature names should *not* contain hyphens. The following names are reserved and should not be used as feature names: ``spkitemid``, ``spkitemlab``, ``itemType``, ``r1``, ``r2``, ``score``, ``sc``, ``sc1``, and ``adj``.
+The name of the feature. This must match the feature name in the training and evaluation ``.csv`` files exactly, including capitalization. Feature names should *not* contain hyphens. The following names are reserved and should not be used as feature names: ``spkitemid``, ``spkitemlab``, ``itemType``, ``r1``, ``r2``, ``score``, ``sc``, ``sc1``, and ``adj``. In addition, any column names provided as values for  ``id_column``, ``train_label_column``, ``test_label_column``, ``length_column``, ``candidate_column``, and ``subgroups`` can also not be used as feature names.
 
 transform
 """""""""
 A transformation that should be applied to the feature values before using it in the model. Possible values are:
 
-    .. hlist::
+    * ``raw``: no transformation, use original feature value
+    * ``org``: same as raw
+    * ``inv``: 1/x
+    * ``sqrt``: square root
+    * ``addOneInv``: 1/(x+1)
+    * ``addOneLn``: ln(x+1)
 
-    * ``raw`` - no transformation, use original feature value
-    * ``org`` - same as raw
-    * ``inv`` - 1/x
-    * ``sqrt`` - square root
-    * ``addOneInv`` - 1/(x+1)
-    * ``addOneLn`` - ln(x+1)
-
-Note that ``rsmtool`` will return an error if the values in the data do not allow the supplied transformation (for example, if ``inv`` is applied to a feature which has 0 values). If you really want to use the tranformation, you must pre-process your training and evaluation ``.csv`` files to remove the problematic cases.
+Note that ``rsmtool`` will raise an exception if the values in the data do not allow the supplied transformation (for example, if ``inv`` is applied to a feature which has 0 values). If you really want to use the tranformation, you must pre-process your training and evaluation ``.csv`` files to remove the problematic cases.
 
 sign
 """"
@@ -62,5 +66,40 @@ To ensure that this is working as expected, you can check the sign of correlatio
 
 Automatic Feature selection
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
 No feature file is necessary for models with automatic feature selection. You simply need to pick one of the built-in models that performs :ref:`automatic feature selection <automatic_feature_selection_models>`.
+
+.. _subset_feature_selection:
+
+Subset-based Feature Selection
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+For more advanced users, ``rsmtool`` offers the ability to define feature subsets and then select groups of features by simply specifying the pre-defined subset to use.
+
[…]
