### 2016-07-22 e020226
[ci skip] Updates to docs
- Fix syntax errors in rsmpredict rst files.
- Update `conf.py` with exclusions.
- Update bookmarks in rsmtool rst files to be more specific.
