+++ doc/config_rsmpredict.rst
@@ -1,21 +1,30 @@
 Experiment configuration file
 """""""""""""""""""""""""""""
-
 This is a file in ``.json`` format that provides overall configuration options for an ``rsmpredict`` experiment. An example configuration file can be found here.
 
 There are three required fields and the rest are all optional.
 
+experiment_dir
+~~~~~~~~~~~~~~
+The path to the directory containing ``rsmtool`` model to use for generating predictions. This directory must contain a sub-directory called ``output`` with the model files, feature pre-processing parameters, and score post-processing parameters. The path can be absolute or relative to the location of configuration file.
+
 experiment_id
 ~~~~~~~~~~~~~
-The original ``experiment_id`` used to create the model files that you will be using for generating predictions. If you do not know the ``experiment_id``, you can find it by looking at the shared first part of the model file names. 
-
-experiment_dir
-~~~~~~~~~~~~~~~
-The path to the directory which contains the outputs of the original `rsmtool` experiment. This directory must contain a directory called ``output`` with the model files, feature pre-processing parameters and score pre-processing parameters. The path can be absolute or relative to the location of config file.
+The ``experiment_id`` used to create the ``rsmtool`` model files being used for generating predictions. If you do not know the ``experiment_id``, you can find it by looking at the prefix of the ``.model`` file under the ``output`` directory.
 
 input_feature_file
-~~~~~~~~~~~~~~~~
-the path to the file with raw feature values that will be used to generate predictions in .csv format. Each row should correspond to a single response and contain feature values for this response. In addition, there should be a column with a unique identifier (ID) for each response. The path can be absolute or relative to the location of config file. Note that the feature names must be the same as used in the original ``rsmtool`` experiment. 
+~~~~~~~~~~~~~~~~~~
+The path to the ``.csv`` file with the raw feature values that will be used for generating predictions. Each row should correspond to a single response and contain feature values for this response. In addition, there should be a column with a unique identifier (ID) for each response. The path can be absolute or relative to the location of config file. Note that the feature names *must* be the same as used in the original ``rsmtool`` experiment.
+
+
+.. note::
+
+    ``rsmpredict`` will only generate predictions for responses in this file that have numeric values for the features included in the ``rsmtool`` model.
+
+
+.. seealso::
+
+    ``rsmpredict`` does not require human scores for the new data since it does not evaluate the generated predictions. If you do have the human scores and want to evaluate the new predictions, you can use the :ref:`rsmeval <usage_rsmeval>` command-line utility.
 
 
 id_column *(Optional)*
@@ -23,19 +32,25 @@ id_column *(Optional)*
 
 The name of the column containing the response IDs. Defaults to ``spkitemid``, i.e., if this is not specified, ``rsmpredict`` will look for a column called ``spkitemid`` in the prediction file.
 
+There are several other options in the configuration file that, while not directly used by ``rsmpredict``, can simply be passed through from the input features file to the output predictions file. This can be particularly useful if you want to subsequently run :ref:`rsmeval <usage_rsmeval>` to evaluate the generated predictions.
 
-Columns that can be added to the predictions (optional)
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+candidate_column *(Optional)*
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+The name for the column containing unique candidate IDs. This column will be named ``candidate`` in the output file with predictions.
 
-The following columns are not used by ``rsmpredict``. You can specify in the config file so that they are included into the predictions file for subsequent use with ``rsmeval``. Different fields are used for different types of columns for consistency with other tools.
[…]
