+++ doc/config_rsmtool.rst
@@ -1,4 +1,4 @@
-.. _config_file:
+.. _config_file_rsmtool:
 
 Experiment configuration file
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
@@ -27,7 +27,7 @@ description *(Optional)*
 """"""""""""""""""""""""
 A brief description of the experiment. This will be included in the report. The description can contain spaces and punctuation. This is blank by default.
 
-.. _feature_file:
+.. _feature_file_rsmtool:
 
 features *(Optional)*
 """""""""""""""""""""
@@ -41,13 +41,13 @@ id_column *(Optional)*
 """"""""""""""""""""""
 The name of the column containing the response IDs. Defaults to ``spkitemid``, i.e., if this is not specified, ``rsmtool`` will look for a column called ``spkitemid`` in the training and evaluation files.
 
-.. _train_label_column:
+.. _train_label_column_rsmtool:
 
 train_label_column *(Optional)*
 """""""""""""""""""""""""""""""
 The name for the column containing the human scores in the training data. If set to to ``fake``, fake scores will be generated using randomly sampled integers. This option may be useful if you only need descriptive statistics for the data and do not care about the other analyses. Defaults to ``sc1``.
 
-.. _test_label_column:
+.. _test_label_column_rsmtool:
 
 test_label_column *(Optional)*
 """"""""""""""""""""""""""""""
@@ -55,19 +55,19 @@ The name for the column containing the human scores in the training data. If set
 
 .. note::
 
-    All responses with non-numeric values in either ``train_label_column`` or ``test_label_column`` and/or those with non-numeric values for relevant features will be automatically excluded from model training and evaluation. By default, zero scores in either ``train_label_column`` or ``test_label_column`` will also be excluded. See :ref:`exclude_zero_scores` if you want to keep responses with zero scores. 
+    All responses with non-numeric values in either ``train_label_column`` or ``test_label_column`` and/or those with non-numeric values for relevant features will be automatically excluded from model training and evaluation. By default, zero scores in either ``train_label_column`` or ``test_label_column`` will also be excluded. See :ref:`exclude_zero_scores_rsmtool` if you want to keep responses with zero scores.
 
-.. _length_column:
+.. _length_column_rsmtool:
 
 length_column *(Optional)*
 """"""""""""""""""""""""""
-The name for the optional column in the training and evaluation data containing response length. If specified, length is included in the inter-feature and partial correlation analyses. Note that this field *should not* be specified if you want to use the length column as an actual feature in the model. In the latter scenario, the length column will automatically be included in the analyses, like any other feature. If you specify ``length_column`` *and* include the same column name as  a feature in the :ref:`feature file <feature_file>`, ``rsmtool`` will ignore the ``length_column`` setting. In addition, if ``length_column`` has missing values or if its standard deviation is 0 (both somewhat unlikely scenarios), ``rsmtool`` will *not* include any length-based analyses in the report.
+The name for the optional column in the training and evaluation data containing response length. If specified, length is included in the inter-feature and partial correlation analyses. Note that this field *should not* be specified if you want to use the length column as an actual feature in the model. In the latter scenario, the length column will automatically be included in the analyses, like any other feature. If you specify ``length_column`` *and* include the same column name as  a feature in the :ref:`feature file <feature_file_rsmtool>`, ``rsmtool`` will ignore the ``length_column`` setting. In addition, if ``length_column`` has missing values or if its standard deviation is 0 (both somewhat unlikely scenarios), ``rsmtool`` will *not* include any length-based analyses in the report.
 
 second_human_score_column *(Optional)*
 """"""""""""""""""""""""""""""""""""""
[…]
