+++ doc/pipeline.rst
@@ -11,7 +11,7 @@ As its primary input, RSMTool takes a CSV feature file with numeric, non-sparse 
 
 The primary output of RSMTool is a comprehensive, customizable HTML statistical report that contains the multiple analyses required for a comprehensive evaluation of an automated scoring model including feature descriptives, subgroup comparisons, model statistics, as well as several different evaluation measures illustrating model efficacy. Details about these various analyses are provided in a separate `technical paper <https://github.com/EducationalTestingService/rsmtool/raw/master/doc/rsmtool.pdf>`_.
 
-In addition to the HTML report, RSMTool also saves the outputs of all of the performed analyses as CSV files.
+In addition to the HTML report, RSMTool also saves the intermediate outputs of all of the performed analyses as CSV files.
 
 
 
