### 2016-07-22 c91cf2c
[ci skip] more fiddling with documentation
- Fix typos in rsmeval rst files and some reorganization
- Add skeleton for usage_rsmcompare.rst
- Some minor edits here and there.
