+++ doc/intermediate_files_rsmeval.rst
@@ -2,8 +2,7 @@
 
 Intermediate CSV files
 """"""""""""""""""""""
-
-Although the primary output of RSMEval is an HTML report, we also want the user to be able to conduct additional analyses outside of RSMEval. To this end, all of the tables produced in an RSMEval experiment report are saved as ``.csv`` files in the ``output`` directory. The following sections describe all of the intermediate files that are produced.
+Although the primary output of ``rsmeval`` is an HTML report, we also want the user to be able to conduct additional analyses outside of ``rsmeval``. To this end, all of the tables produced in an the experiment report are saved as ``.csv`` files in the ``output`` directory. The following sections describe all of the intermediate files that are produced.
 
 .. note::
 
@@ -18,44 +17,47 @@ Predictions
 ~~~~~~~~~~~
 filename: ``pred_processed.csv``
 
-This file contains the processed predicted scores for the evaluation set: the original predictions are processed to produced truncated and rounded scores as well as scaled scores if requested. 
-
+This file contains the post-processed predicted scores: the predictions from the model are truncated, rounded, and re-scaled (if requested).
 
 Flagged responses
 ~~~~~~~~~~~~~~~~~
 filename: ``test_responses_with_excluded_flags.csv``
 
-This file contains all of the rows in predictions file that were filtered out based on conditions specified in :ref:`flag_column <flag_column_eval>`.  Note that if the original predictions ``.csv`` files contained columns with internal names such as ``sc1`` but these columns were not actually used by ``rsmeval``, these columns will also be included into these files but their names will be changed to ``##name##`` (e.g. ``##sc1##``).
+This file contains all of the rows in input predictions file that were filtered out based on conditions specified in :ref:`flag_column <flag_column_eval>`.
+
+.. note::
+
+    If the predictions file contained columns with internal names such as ``sc1`` that were not actually used by ``rsmeval``, they will still be included in these files but their names will be changed to ``##name##`` (e.g. ``##sc1##``).
 
 Excluded responses
 ~~~~~~~~~~~~~~~~~~
 filename: ``test_excluded_responses.csv``
 
-This file contains all of the rows in the predictions file that were filtered out because of non-numeric or zero scores. 
+This file contains all of the rows in the predictions file that were filtered out because of non-numeric or zero scores.
 
 Response metadata
-~~~~~~~~~~~~~~~~"
+~~~~~~~~~~~~~~~~~
 filename:  ``test_metadata.csv``
 
-This file contains the metadata columns (``id_column``,  ``subgroups`` if provided) for all responses that were used for the analysis. 
+This file contains the metadata columns (``id_column``,  ``subgroups`` if provided) for all rows in the predictions file that used in the evaluation.
 
 Unused columns
 ~~~~~~~~~~~~~~
 filename: ``test_other_columns.csv``
 
-This file contains all of the the columns from the predictions file that are not present in the ``*_pred_processed.csv`` and ``*_metadata.csv`` files. They only include the rows that were not filtered out. Note that if the original``.csv`` files contained columns with internal names such as ``sc1`` but these columns were not actually used by ``rsmeval``, these columns will also be included into these files but their names will be changed to ``##name##`` (e.g. ``##sc1##``).
+This file contains all of the the columns from the input predictions file that are not present in the ``*_pred_processed.csv`` and ``*_metadata.csv`` files. They only include the rows that were not filtered out.
 
-Response length
-~~~~~~~~~~~~~~"
-filename: ``train_response_lengths.csv``
+.. note::
 
-If `length_column` is specified, then this file contains the values from that column for the training data under a column called ``length`` with the response IDs under the ``spkitemid`` column.
[…]
