+++ doc/usage_rsmcompare.rst
@@ -1,0 +1,2 @@
+``rsmcompare`` - Compare scoring models
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
