+++ doc/advanced_usage.rst
@@ -1,8 +1,7 @@
 Advanced Uses of RSMTool
 ------------------------
 
-In addition to training and evaluating regression-based scoring models, RSMTool provides three other command-line scripts for more advanced users.
-
+In addition to providing the ``rsmtool`` utility training and evaluating regression-based scoring models, the RSMTool package also provides three other command-line utilities for more advanced users.
 
 .. include:: usage_rsmeval.rst
 
