### 2016-07-22 7d73c3e
[ci skip] Link aux tools to language in who.rst
- Update conf.py to exclude `usage_rsmcompare.rst`
