+++ doc/usage_rsmcompare.rst
@@ -1,2 +1,4 @@
+.. _usage_rsmcompare:
+
 ``rsmcompare`` - Compare scoring models
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
