### 2016-07-22 afe4273
[ci skip] Change tree depth in main file to 2
