+++ doc/index.rst
@@ -28,7 +28,7 @@ Documentation
 =============
 
 .. toctree::
-   :maxdepth: 3
+   :maxdepth: 2
 
    who
    pipeline
