### 2016-07-22 d10d6c1
Suppress unnecessary warnings during testing
- Add information to README about how to suppress warnings.
