+++ tests/test_analysis.py
@@ -1,6 +1,8 @@
+import warnings
+
 import pandas as pd
 
-from nose.tools import (assert_almost_equal, assert_equal)
+from nose.tools import (assert_almost_equal, assert_equal, with_setup)
 from numpy.random import RandomState
 
 from rsmtool.analysis import (correlation_helper,
@@ -28,14 +30,26 @@ def test_correlation_helper():
     assert_equal(retval[0].isnull().values.sum(), 0)
     assert_equal(retval[1].isnull().values.sum(), 0)
 
+def setup_correlation_one_row():
+    """
+    Temporarily disable runtime warnings since they are
+    expected for this particular test function with
+    a single row of data/
+    """
+    warnings.filterwarnings('ignore', category=RuntimeWarning)
 
+def teardown_correlation_one_row():
+    """
+    Re-enable warnings now that the test function has run.
+    """
+    warnings.filterwarnings('always', category=RuntimeWarning)
 
+@with_setup(setup_correlation_one_row, teardown_correlation_one_row)
 def test_that_correlation_helper_works_for_data_with_one_row():
     # this should return two data frames with nans
     retval = correlation_helper(df_features[:1], 'sc1', 'group')
     assert_equal(retval[0].isnull().values.sum(), 3)
     assert_equal(retval[1].isnull().values.sum(), 3)
-
 
 
 def test_that_correlation_helper_works_for_data_with_two_rows():
