### 2016-07-22 55634f8
Suppress unnecessary warnings during testing
- Add information to README about how to suppress warnings.
