+++ doc/feature_selection_rsmtool.rst
@@ -1,9 +1,14 @@
 .. _feature_selection:
 
-Selecting Features
-------------------
+Selecting Feature Columns
+-------------------------
 
-By default, ``rsmtool`` will use all of the features provided in the training and evaluation ``.csv`` files. However, it is possible for you to either manually choose a subset of features or to have ``rsmtool`` perform automatic feature selection.
+By default, ``rsmtool`` will use all columns included in the training and evaluation ``.csv`` files as features. The only exception are the columns specified in the :ref:`rsmtool configuration file <flag_column_rsmtool>` as containing ``id``, ``train_label``, ``subgroups`` or other information. However, it is possible for you to define a subset of columns to be used as features. You can then either use all these features in the final model or use one of the built-in models that perform :ref:`automatic feature selection <automatic_feature_selection_models>`.
+
+.. note::
+
+    For all feature selection methods ``rsmtool`` will pre-process the data to remove all responses with non-numeric feature values for any of the features before trying to train a model. If your data set includes a column containing character (string) values and you do not specify this column as an ``id`` column, ``candidate`` column or ``subgroup`` column, ``rsmtool`` will assume that this column contains a feature and will filter out all responses in the data set. 
+
 
 .. note::
 
@@ -11,9 +16,9 @@ By default, ``rsmtool`` will use all of the features provided in the training an
 
 .. _manual_feature_selection:
 
-Manual Feature selection
-^^^^^^^^^^^^^^^^^^^^^^^^
-To manually select a subset of features, you must provide a ``.json`` file specifying information about what features should be included in the final scoring model. For additional flexibility, the same file also allows you to describe transformations to be applied to the raw features before being included in the model.
+Defining Feature Columns using ``.json`` file
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+To manually select a subset of columns to be used as features, you can provide a ``.json`` file specifying information about what columns should be included as features in the final scoring model. For additional flexibility, the same file also allows you to describe transformations to be applied to the raw features before being included in the model.
 
 Here's an example of what this file looks like.
 
@@ -64,35 +69,43 @@ When determining the sign, you should take into account the correlation between 
 
 To ensure that this is working as expected, you can check the sign of correlations for both raw and processed features in the final report.
 
-Automatic Feature selection
-^^^^^^^^^^^^^^^^^^^^^^^^^^^
-No feature file is necessary for models with automatic feature selection. You simply need to pick one of the built-in models that performs :ref:`automatic feature selection <automatic_feature_selection_models>`.
-
 .. _subset_feature_selection:
 
-Subset-based Feature Selection
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-For more advanced users, ``rsmtool`` offers the ability to define feature subsets and then select groups of features by simply specifying the pre-defined subset to use.
+Defining Feature Columns using ``.csv`` file
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+For more advanced users, ``rsmtool`` offers the ability to define feature subsets in a ``.csv`` file and then select groups of columns to be used as features by simply specifying the pre-defined subset to use.
 
-To define feature subsets, you must provide a master ``.csv`` file which lists *all* features that should be used for feature selection. Each feature name should be listed under a column named ``Feature``. Then each subset is an additional column with the value of either ``0`` (denoting that the feature does *not* belong to the subset named by that column) or ``1`` (denoting that the feature does belong to the subset).
[…]
