### 2016-07-22 2df24f8
Suppress warnings on CircleCI too.
