### 2016-07-22 ac1c9c2
Changed pronouns in the example and added links to external tools
