### 2016-07-22 fdbfdac
Suppressing warnings properly
- Fix broken test that breaks CI build.
