### 2016-07-22 7f7f927
Suppressing warnings properly
- Fix broken test that breaks CI build.
