### 2016-07-22 3377996
[ci skip] documentation for rsmcompare
