### 2016-07-22 7cb7751
More fiddling with warnings.
