+++ rsmtool/__init__.py
@@ -3,6 +3,7 @@
 :organization: ETS
 """
 
+import os
 import re
 import warnings
 
@@ -18,7 +19,9 @@ from .rsmtool import run_experiment
 __all__ = ['run_experiment']
 
 # Make sure that DeprecationWarnings are always shown
-# within this package.
-warnings.filterwarnings('always',
-                        category=DeprecationWarning,
-                        module='^{0}\.'.format(re.escape(__name__)))
+# within this package unless we are in test mode in
+# which case do not enable them by default.
+if 'TEST_MODE' not in os.environ:
+    warnings.filterwarnings('always',
+                            category=DeprecationWarning,
+                            module='^{0}\.'.format(re.escape(__name__)))
