### 2016-07-22 0c824cc
More fiddling with warnings.
