### 2016-07-22 1e2077d
Rename feature selection to column selection.
