+++ doc/column_selection_rsmtool.rst
@@ -1,0 +1,128 @@
+.. _column_selection_rsmtool:
+
+Selecting Feature Columns
+-------------------------
+
+By default, ``rsmtool`` will use all columns included in the training and evaluation ``.csv`` files as features. The only exception are any columns explicitly identified in the configuration file as containing non-feature information (e.g., :ref:`id_column <id_column_rsmtool>`, :ref:`train_label_column <train_label_column_rsmtool>`, :ref:`test_label_column <test_label_column_rsmtool>`, etc.)
+
+However, it is possible for you to choose which columns in the data should be used as features. There are two ways to do so:
+
+    1. **Fine-grained column selection**: In this method, you manually create a list of the columns that you wish to use as features for an ``rsmtool`` experiment. See :ref:`fine-grained selection <json_column_selection>` for more details.
+
+    2. **Subset-based column selection**: In this method, you can pre-define subsets of features and then select entire subsets at a time for any ``rsmtool`` experiment. See :ref:`subset-based selection <subset_column_selection>` for more details.
+
+.. warning::
+
+    For both column selection methods, ``rsmtool`` will filter the training and evaluation data to remove any responses with non-numeric values in any of the selected columns before training the model. If your data includes a column containing string values and you do *not* specify this column as the ``id_column`` or the ``candidate_column``  or a ``subgroup`` column, ``rsmtool`` might filter out *all* the responses in the data.
+
+.. note::
+
+        For all column selection methods, the final set of features will be saved in the ``feature`` folder in the experiment output directory.
+
+
+.. _json_column_selection:
+
+Fine-grained column selection
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+To manually select a subset of columns to be used as features, you can provide a ``.json`` file specifying information about what columns should be included as features in the final scoring model. For additional flexibility, the same file also allows you to describe transformations to be applied to the raw features before being included in the model.
+
+Here's an example of what this file looks like.
+
+
+.. code-block:: javascript
+
+    {
+        "features": [{
+                         "feature": "feature1",
+                         "transform": "raw",
+                         "sign": 1,
+                     },
+                     {
+                         "feature": "feature2",
+                         "transform": "inv",
+                         "sign": -1,
+                     },
+                     ...
+                    ]
+    }
+
+
+There are three required fields.
+
+feature
+"""""""
+The name of the feature. This must match the feature name in the training and evaluation ``.csv`` files exactly, including capitalization. Feature names should *not* contain hyphens. The following names are reserved and should not be used as feature names: ``spkitemid``, ``spkitemlab``, ``itemType``, ``r1``, ``r2``, ``score``, ``sc``, ``sc1``, and ``adj``. In addition, any column names provided as values for  ``id_column``, ``train_label_column``, ``test_label_column``, ``length_column``, ``candidate_column``, and ``subgroups`` can also not be used as feature names.
+
+transform
+"""""""""
+A transformation that should be applied to the feature values before using it in the model. Possible values are:
+
+    * ``raw``: no transformation, use original feature value
+    * ``org``: same as raw
+    * ``inv``: 1/x
+    * ``sqrt``: square root
+    * ``addOneInv``: 1/(x+1)
+    * ``addOneLn``: ln(x+1)
+
+Note that ``rsmtool`` will raise an exception if the values in the data do not allow the supplied transformation (for example, if ``inv`` is applied to a feature which has 0 values). If you really want to use the tranformation, you must pre-process your training and evaluation ``.csv`` files to remove the problematic cases.
+
+sign
+""""
+
[…]
