+++ doc/feature_selection_rsmtool.rst
@@ -1,122 +1,0 @@
-.. _feature_selection:
-
-Selecting Feature Columns
--------------------------
-
-By default, ``rsmtool`` will use all columns included in the training and evaluation ``.csv`` files as features. The only exception are the columns specified in the :ref:`rsmtool configuration file <flag_column_rsmtool>` as containing ``id``, ``train_label``, ``subgroups`` or other information. However, it is possible for you to define a subset of columns to be used as features. You can then either use all these features in the final model or use one of the built-in models that perform :ref:`automatic feature selection <automatic_feature_selection_models>`.
-
-.. note::
-
-    For all feature selection methods ``rsmtool`` will pre-process the data to remove all responses with non-numeric feature values for any of the features before trying to train a model. If your data set includes a column containing character (string) values and you do not specify this column as an ``id`` column, ``candidate`` column or ``subgroup`` column, ``rsmtool`` will assume that this column contains a feature and will filter out all responses in the data set. 
-
-
-.. note::
-
-    For all feature selection methods, the final set of features will be saved in the ``feature`` folder in the experiment output directory.
-
-.. _manual_feature_selection:
-
-Defining Feature Columns using ``.json`` file
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-To manually select a subset of columns to be used as features, you can provide a ``.json`` file specifying information about what columns should be included as features in the final scoring model. For additional flexibility, the same file also allows you to describe transformations to be applied to the raw features before being included in the model.
-
-Here's an example of what this file looks like.
-
-
-.. code-block:: javascript
-
-    {
-        "features": [{
-                         "feature": "feature1",
-                         "transform": "raw",
-                         "sign": 1,
-                     },
-                     {
-                         "feature": "feature2",
-                         "transform": "inv",
-                         "sign": -1,
-                     },
-                     ...
-                    ]
-    }
-
-
-There are three required fields.
-
-feature
-"""""""
-The name of the feature. This must match the feature name in the training and evaluation ``.csv`` files exactly, including capitalization. Feature names should *not* contain hyphens. The following names are reserved and should not be used as feature names: ``spkitemid``, ``spkitemlab``, ``itemType``, ``r1``, ``r2``, ``score``, ``sc``, ``sc1``, and ``adj``. In addition, any column names provided as values for  ``id_column``, ``train_label_column``, ``test_label_column``, ``length_column``, ``candidate_column``, and ``subgroups`` can also not be used as feature names.
-
-transform
-"""""""""
-A transformation that should be applied to the feature values before using it in the model. Possible values are:
-
-    * ``raw``: no transformation, use original feature value
-    * ``org``: same as raw
-    * ``inv``: 1/x
-    * ``sqrt``: square root
-    * ``addOneInv``: 1/(x+1)
-    * ``addOneLn``: ln(x+1)
-
-Note that ``rsmtool`` will raise an exception if the values in the data do not allow the supplied transformation (for example, if ``inv`` is applied to a feature which has 0 values). If you really want to use the tranformation, you must pre-process your training and evaluation ``.csv`` files to remove the problematic cases.
-
-sign
-""""
-
-After transformation, each feature value will be multiplied by this number. This field is usually set to ``1`` or ``-1`` depending on the expected sign of the correlation between transformed feature and human score to ensure that all features in the final models have positive correlation with the score.
-
[…]
