### 2016-07-22 5a23294
[ci skip] Re-organize the column selection information
