+++ doc/column_selection_rsmtool.rst
@@ -5,11 +5,15 @@ Selecting Feature Columns
 
 By default, ``rsmtool`` will use all columns included in the training and evaluation ``.csv`` files as features. The only exception are any columns explicitly identified in the configuration file as containing non-feature information (e.g., :ref:`id_column <id_column_rsmtool>`, :ref:`train_label_column <train_label_column_rsmtool>`, :ref:`test_label_column <test_label_column_rsmtool>`, etc.)
 
-However, it is possible for you to choose which columns in the data should be used as features. There are two ways to do so:
+However, there are certain scenarios in which it is useful to choose specific columns in the data to be used as features. For example, let's say that you have a large number of very different features and you want to use a different subset of features to score different types of questions on a test. In this case, the ability to easily choose the desired features for any ``rsmtool`` experiment becomes quite important. The alternative of manually pre-processing the data to remove the features you don't need is quite cumbersome.
+
+There are two ways to select specific columns in the data as features:
 
     1. **Fine-grained column selection**: In this method, you manually create a list of the columns that you wish to use as features for an ``rsmtool`` experiment. See :ref:`fine-grained selection <json_column_selection>` for more details.
 
     2. **Subset-based column selection**: In this method, you can pre-define subsets of features and then select entire subsets at a time for any ``rsmtool`` experiment. See :ref:`subset-based selection <subset_column_selection>` for more details.
+
+While fine-grained column selection is better for a single experiment, subset-based selection is more convenient when you need to run several experiments with somewhat different subsets of features.
 
 .. warning::
 
@@ -24,9 +28,9 @@ However, it is possible for you to choose which columns in the data should be us
 
 Fine-grained column selection
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-To manually select a subset of columns to be used as features, you can provide a ``.json`` file specifying information about what columns should be included as features in the final scoring model. For additional flexibility, the same file also allows you to describe transformations to be applied to the raw features before being included in the model.
+To manually select columns to be used as features, you can provide a ``.json`` file specifying the names of the columns that should be included in the final scoring model. For additional flexibility, the same file also allows you to describe transformations to be applied to the the values in these feature columns before being used in the model.
 
-Here's an example of what this file looks like.
+Here's an example of what such a file might look like.
 
 
 .. code-block:: javascript
@@ -51,25 +55,24 @@ There are three required fields.
 
 feature
 """""""
-The name of the feature. This must match the feature name in the training and evaluation ``.csv`` files exactly, including capitalization. Feature names should *not* contain hyphens. The following names are reserved and should not be used as feature names: ``spkitemid``, ``spkitemlab``, ``itemType``, ``r1``, ``r2``, ``score``, ``sc``, ``sc1``, and ``adj``. In addition, any column names provided as values for  ``id_column``, ``train_label_column``, ``test_label_column``, ``length_column``, ``candidate_column``, and ``subgroups`` can also not be used as feature names.
+The exact name of the column in the training and evaluation ``.csv`` files, including capitalization. Column names cannot contain hyphens. The following strings are reserved and cannot not be used as feature column names: ``spkitemid``, ``spkitemlab``, ``itemType``, ``r1``, ``r2``, ``score``, ``sc``, ``sc1``, and ``adj``. In addition, any column names provided as values for  ``id_column``, ``train_label_column``, ``test_label_column``, ``length_column``, ``candidate_column``, and ``subgroups`` may also not be used as feature column names.
 
 transform
 """""""""
-A transformation that should be applied to the feature values before using it in the model. Possible values are:
+A transformation that should be applied to the column values before using it in the model. Possible values are:
 
[…]
