### 2016-07-22 29f7895
Make it really simple
- python provides a really nice way to temporarily suppress warnings using the `catch_warnings` context manager so we should use that.
