### 2016-07-22 d28f98e
Make it really simple
- python provides a really nice way to temporarily suppress warnings using the `catch_warnings` context manager so we should use that.
