+++ rsmtool/__init__.py
@@ -3,7 +3,6 @@
 :organization: ETS
 """
 
-import os
 import re
 import warnings
 
@@ -21,7 +20,6 @@ __all__ = ['run_experiment']
 # Make sure that DeprecationWarnings are always shown
 # within this package unless we are in test mode in
 # which case do not enable them by default.
-if 'TEST_MODE' not in os.environ:
-    warnings.filterwarnings('always',
-                            category=DeprecationWarning,
-                            module='^{0}\.'.format(re.escape(__name__)))
+warnings.filterwarnings('always',
+                        category=DeprecationWarning,
+                        module='^{0}\.'.format(re.escape(__name__)))
