### 2016-07-22 cdb5e2f
Suppress runtime warnings in test_preprocess.py.
