### 2016-07-24 2028f4d
[ci skip] Lots of changes
- Add `config_rsmcompare.rst`.
- Add `custom_notebooks.rst`.
- Some minor changes in `usage_rsmcompare.rst`.
