### 2016-07-24 b37bb41
[ci skip] Some language change and new links.
