### 2016-07-25 1f49d3b
Change the last set of notes in custom_notebooks.rst to footnotes.
