+++ doc/custom_notebooks.rst
@@ -126,11 +126,11 @@ figure_dir
 ^^^^^^^^^^
 The ``figure`` sub-directory under the experiment output directory that contains all of the generated SVG and PNG figures.
 
-.. note::
+.. rubric:: Footnotes
 
-    1. All dataframes apart from ``df_train_orig`` and  ``df_test_orig`` contain an ``spkitemid`` column which contains the unique response IDs.
+.. [#] All dataframes apart from ``df_train_orig`` and  ``df_test_orig`` contain an ``spkitemid`` column which contains the unique response IDs.
 
-    2. All data frames except the ``df_*_other_columns`` contain an ``sc1`` column which contains the human score for the responses.
+.. [#] All data frames except the ``df_*_other_columns`` contain an ``sc1`` column which contains the human score for the responses.
 
-    3. ``df_train_orig`` and ``df_test_orig`` will contain the response IDs and human scores under columns with the original names, *not* ``spkitemid`` and ``sc1``.
+.. [#] ``df_train_orig`` and ``df_test_orig`` will contain the response IDs and human scores under columns with the original names, *not* ``spkitemid`` and ``sc1``.
 
