+++ doc/custom_notebooks.rst
@@ -4,6 +4,8 @@ Writing custom RSMTool sections
 -------------------------------
 RSMTool allows users to include custom Jupyter notebooks in the reports generated by :ref:`rsmtool <usage_rsmtool>`, :ref:`rsmeval <usage_rsmeval>`, and :ref:`rsmcompare <usage_rsmcompare>` using the ``custom_sections`` fields in their configuration files. This can be particularly useful if a researcher wants to include custom analyses specific to their own scoring engine; she can do so while still using the familiar RSMTool pipeline.
 
+Available variables
+~~~~~~~~~~~~~~~~~~~
 When writing such notebooks, some or all of the python variables below will be available for use in the notebooks.
 
 experiment_id
