### 2016-07-25 18bc095
Minor edits
