### 2016-07-25 1aaeb75
[ci skip] fix minor typos and merge some notes.
