+++ doc/builtin_models.rst
@@ -32,9 +32,6 @@ Models with automatic feature selection
 
 .. note::
 
-    ``NNLR``, ``LassoFixedLambdaThenNNLR``, ``LassoFixedLambda`` and ``PositiveLassoCV`` all have no negative coefficients.
+    1. ``NNLR``, ``LassoFixedLambdaThenNNLR``, ``LassoFixedLambda`` and ``PositiveLassoCV`` all have no negative coefficients.
 
-
-.. note::
-
-        For all feature selection models, the final set of features will be saved in the ``feature`` folder in the experiment output directory.
+    2. For all feature selection models, the final set of features will be saved in the ``feature`` folder in the experiment output directory.
