### 2016-07-25 1cedcd5
[ci skip] edits to rsmcompare; added consistency to the list of sections for all tools; re-ordered section descriptions based on default order
