+++ doc/config_rsmcompare.rst
@@ -17,11 +17,11 @@ An identifier for the experiment with the "new" model (e.g., the model with new 
 
 experiment_dir_old
 ~~~~~~~~~~~~~~~~~~
-The directory with the results for the "baseline" experiment. This directory is the output directory that was used for the experiment and should contain subdirectories ``output`` and ``figure``.
+The directory with the results for the "baseline" experiment. This directory is the output directory that was used for the experiment and should contain subdirectories ``output`` and ``figure`` generated by ``rsmtool`` or ``rsmeval``.
 
 experiment_dir_new
 ~~~~~~~~~~~~~~~~~~
-The directory with the results for the experiment with the new model. This directory is the output directory that was used for the experiment and should contain subdirectories ``output`` and ``figure``.
+The directory with the results for the experiment with the new model. This directory is the output directory that was used for the experiment and should contain subdirectories ``output`` and ``figure`` generated by ``rsmtool`` or ``rsmeval``.
 
 description_old
 ~~~~~~~~~~~~~~~
@@ -39,9 +39,13 @@ use_scaled_predictions_new *(Optional)*
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 Set to ``true`` if the experiment with the new model used scaled machine scores for confusion matrices, score distributions, subgroup analyses, etc. Defaults to ``false``.
 
+.. note::
+
+    For ``rsmtool`` and ``rsmeval``, main evaluations are computed for both scaled and raw scores, but in-depth evaluations such as confusion matrix are only computed for one type of scores as defined in the field ``use_scaled_predictions`` .  ``rsmcompare`` uses the existing outputs and does not perform any additional evaluations. Therefore if this field was set to ``true`` in the original experiment and to ``false`` for ``rsmcompare``, the report will be inconsistent: main evaluations will be displayed for raw scores, while in-depth evaluations will be displayed for scaled scores. 
+
 subgroups *(Optional)*
 ~~~~~~~~~~~~~~~~~~~~~~
-A list of column names indicating grouping variables used for generating analyses specific to each of those defined subgroups.
+A list of column names indicating grouping variables used for generating analyses specific to each of those defined subgroups.For example, ``["prompt, gender, native_language, test_country"]``. 
 
 .. note::
 
@@ -55,9 +59,10 @@ RSMTool provides pre-defined sections for ``rsmcompare`` (listed below) and, by 
 
     - ``feature_descriptives``: Compares the descriptive statistics for all raw feature values included in the model:
 
-        - a table showing mean, standard deviation, min, max, correlation with human score etc.;
-        - a table with percentiles and outliers; and
-        - a barplot showing he number of truncated outliers for each feature.
+        - a table showing mean, standard deviation, skewness and kurtosis;
+        - a table showing the number of truncated outliers for each feature; and
+        - a table with percentiles and outliers;
+        
 
     - ``features_by_group``: Shows boxplots for both experiments with distributions of raw feature values by each of the :ref:`subgroups <subgroups_rsmtool>` specified in the configuration file.
 
@@ -65,29 +70,36 @@ RSMTool provides pre-defined sections for ``rsmcompare`` (listed below) and, by 
 
         - histograms showing the distributions of preprocessed features values;
         - the correlation matrix between all features and the human score;
-        - a barplot showing marginal and partial correlations between all features and the human score, and, optionally, response length if :ref:`length_column <length_column_rsmtool>` is specified in the config file.
+        - a table showing marginal correlations between all features and the human score; and
+        - a table showing partial correlations between all features and the human score.
 
     - ``preprocessed_features_by_group``: Compares analyses of preprocessed features by subgroups: marginal and partial correlations between each feature and human score for each subgroup.
 
[…]
