### 2016-07-25 8ab2f02
[ci skip] fix some typos and convert note to warnings
