### 2016-07-25 68afcb2
[ci skip] Add links to example config files.
- Also rename `config.json` to `config_rsmtool.json` for consistency.
