+++ doc/config_rsmpredict.rst
@@ -1,6 +1,6 @@
 Experiment configuration file
 """""""""""""""""""""""""""""
-This is a file in ``.json`` format that provides overall configuration options for an ``rsmpredict`` experiment. An example configuration file can be found here.
+This is a file in ``.json`` format that provides overall configuration options for an ``rsmpredict`` experiment. An example configuration file can be found `here <https://github.com/EducationalTestingService/rsmtool/blob/master/example/rsmpredict/config_rsmpredict.json>`_.
 
 There are three required fields and the rest are all optional.
 
