### 2016-07-25 84b725e
Description of pre-processing parameters:
