### 2016-07-25 0a513b2
Show command-line usage for all four scripts.
