### 2016-07-25 f0772bc
[ci skip] beginnings of API documentation.
