+++ rsmtool/__init__.py
@@ -14,8 +14,10 @@ except ImportError:
 else:
     HAS_RSMEXTRA = True
 
+from .analysis import compute_basic_descriptives
 from .rsmtool import run_experiment
-__all__ = ['run_experiment']
+
+__all__ = ['run_experiment', 'compute_basic_descriptives']
 
 # Make sure that DeprecationWarnings are always shown
 # within this package
