### 2016-07-25 3e5e974
More api docs from analysis.py
