### 2016-07-25 9c26966
[ci skip] Add functions from `preprocess.py`
- And figure out MathJax support.
