### 2016-07-25 be6d9bd
input.py
