+++ doc/api.rst
@@ -25,3 +25,7 @@ From :py:mod:`~rsmtool:create_features.py` Module
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 .. autofunction:: rsmtool.generate_default_specs
 .. autofunction:: rsmtool.find_feature_transformation
+
+From :py:mod:`~rsmtool:input.py` Module
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+.. autofunction:: rsmtool.parse_json_with_comments
