+++ doc/api.rst
@@ -29,3 +29,15 @@ From :py:mod:`~rsmtool:create_features.py` Module
 From :py:mod:`~rsmtool:input.py` Module
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 .. autofunction:: rsmtool.parse_json_with_comments
+
+From :py:mod:`~rsmtool:model.py` Module
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+.. autofunction:: rsmtool.model_fit_to_dataframe
+.. autofunction:: rsmtool.ols_coefficients_to_dataframe
+.. autofunction:: rsmtool.skll_learner_params_to_dataframe
+..autofunction:: rsmtool.train_builtin_model
+
+
+From :py:mod:`~rsmtool:predict.py` Module
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+.. autofunction:: rsmtool.predict_with_model
