### 2016-07-25 85ae25c
[ci skip] More API docs.
