+++ doc/index.rst
@@ -37,6 +37,7 @@ Documentation
    advanced_usage
    custom_notebooks
    api
+   utilities
    contributing
 
 Indices and tables
