+++ doc/api.rst
@@ -2,7 +2,7 @@
 
 API Documentation
 -----------------
-The primary method of using RSMTool is via the command-line scripts ``rsmtool``, ``rsmeval``, ``rsmpredict``, and ``rsmcompare``. However, there are certain functions in the ``rsmtool`` API that may also be useful to advanced users.
+The primary method of using RSMTool is via the command-line scripts :ref:`rsmtool <usage_rsmtool>`, :ref:`rsmeval <usage_rsmeval>`, :ref:`rsmpredict <usage_rsmpredict>`, and :ref:`rsmcompare <usage_rsmcompare>`. However, there are certain functions in the ``rsmtool`` API that may also be useful to advanced users for use directly in their Python code. We document these functions below.
 
 :mod:`rsmtool` Package
 ======================
@@ -29,3 +29,7 @@ From :py:mod:`~rsmtool:create_features.py` Module
 From :py:mod:`~rsmtool:input.py` Module
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 .. autofunction:: rsmtool.parse_json_with_comments
+
+From :py:mod:`~rsmtool.report` Module
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+.. autofunction:: rsmtool.convert_ipynb_to_html
