### 2016-07-25 2555e0e
beginning of report.py
