### 2016-07-25 5c49af2
Fixed formatting typos
