### 2016-07-25 e7ba56c
[ci skip] some tweaks to docstrings.
