### 2016-07-25 c0d76f2
Fix some links and some minor changes.
