+++ doc/pipeline.rst
@@ -8,47 +8,48 @@ The following figure gives an overview of the RSMTool pipeline:
    :align: center
 
 
-As its primary input, RSMTool takes a CSV feature file with numeric, non-sparse features and a human score as input, :ref:`preprocesses <preprocessing_steps>` them and lets you try several different regression models (including Ridge, SVR, AdaBoost, and Random Forests) to try and predict the human score from the features.
-The model is then applied to generate scores for the new data set preprocessed using the same :ref:`preprocessing parameters <preprocessing_parameters>`. In addition to ``raw`` predicted scores, the ``prediction analysis`` component of the RSMTool generates several types of :ref:`post-processed scores <score_postprocessing>` commonly used in automated scoring. 
+As its primary input, RSMTool takes a CSV feature file with numeric, non-sparse features and a human score as input, :ref:`pre-processes <feature_preprocessing>` them and lets you train a regression-based **Scoring Model** to predict the human score from the features. Available regression models include Ridge, SVR, AdaBoost, and Random Forests, among many others.
 
-The primary output of RSMTool is a comprehensive, customizable HTML statistical report that contains the multiple analyses required for a comprehensive evaluation of an automated scoring model including feature descriptives, subgroup comparisons, model statistics, as well as several different evaluation measures illustrating model efficacy. Details about these various analyses are provided in a separate `technical paper <https://github.com/EducationalTestingService/rsmtool/raw/master/doc/rsmtool.pdf>`_.
+This trained model can then be used to generate scores for a held-out evaluation data whose feature values are pre-processed using the same :ref:`Pre-processing Parameters <preprocessing_parameters>`. In addition to the ``raw`` scores predicted by the model, the **Prediction Analysis** component of the pipline generates several additional :ref:`post-processed scores <score_postprocessing>` that are commonly used in automated scoring.
+
+The primary output of RSMTool is a comprehensive, customizable HTML statistical report that contains the multiple analyses required for a comprehensive evaluation of an automated scoring model including descriptive analyses for all features, model analyses, subgroup comparisons, as well as several different evaluation measures illustrating model efficacy. Details of these various analyses are provided in a separate `technical paper <https://github.com/EducationalTestingService/rsmtool/raw/master/doc/rsmtool.pdf>`_.
 
 In addition to the HTML report, RSMTool also saves the intermediate outputs of all of the performed analyses as :ref:`CSV files <intermediate_files_rsmtool>`.
 
-.. _preprocessing_steps:
+.. _feature_preprocessing:
 
-Preprocessing steps
-"""""""""""""""""""
+Feature pre-processing
+""""""""""""""""""""""
 
 Data filtering
 ~~~~~~~~~~~~~~
 
-1. Remove all responses with non-numeric values for at least one of the features (see :ref:`column selection methods <column_selection_rsmtool>` for different ways to select features).
+1. Remove all training and evaluation responses that have non-numeric for any of the features (see :ref:`column selection methods <column_selection_rsmtool>` for different ways to select features).
 
-2. Remove all responses with non-numeric values of :ref:`train_label_column <train_label_column_rsmtool>` or :ref:`test_label_column <test_label_column_rsmtool>`. 
+2. Remove all training and evaluation responses with non-numeric values for human scores.
 
[…]
