### 2016-07-26 ff86c7c
[ci skip] minor changes here and there.
