### 2016-07-26 d626544
[ci skip] Add even more API docs.
