### 2016-07-26 4b62f8e
Added three new functions to API
