+++ rsmtool/rsmeval.py
@@ -49,8 +49,21 @@ from rsmtool.version import __version__
 
 def run_evaluation(config_file, output_dir):
     """
-    Run RSMTool evaluation experiment using the given configuration
-    file and generate all evaluation outputs in the given directory.
+    Run RSMEval experiment using the given configuration
+    file and generate all outputs in the given directory.
+
+    Parameters
+    ----------
+    config_file : str
+        Path to the experiment configuration file.
+    output_dir : str
+        Path to the experiment output directory.
+
+    Raises
+    ------
+    ValueError
+        If any of the required fields are missing or ill-specified.
+
     """
 
     logger = logging.getLogger(__name__)
