### 2016-07-26 a547088
Added three new functions to API
