### 2016-07-26 82d7a6e
[ci skip] Some minor typos.
