### 2016-07-26 187a51b
First batch of issues
