+++ rsmtool/input.py
@@ -32,6 +32,7 @@ if HAS_RSMEXTRA:
     from rsmextra.settings import (default_feature_subset_file,
                                    default_feature_sign)
 
+
 def select_candidates_with_N_or_more_items(df,
                                            N,
                                            candidate_column='candidate'):
@@ -72,9 +73,6 @@ def select_candidates_with_N_or_more_items(df,
 
     return (df_included,
             df_excluded)
-
-
-
 
 
 def locate_custom_sections(custom_report_section_paths, configpath):
@@ -146,7 +144,7 @@ def normalize_json_fields(json_obj):
                           'scale': 'use_scaled_predictions',
                           'feature.subset': 'feature_subset'}
 
-    model_name_mapping = {'empWt':'LinearRegression',
+    model_name_mapping = {'empWt': 'LinearRegression',
                           'eqWt': 'EqualWeightsLR',
                           'empWtBalanced': 'RebalancedLR',
                           'empWtDropNeg': '',
@@ -346,7 +344,7 @@ def validate_and_populate_json_fields(json_obj, context='rsmtool'):
     # and sign has not been specified in the config file
     if HAS_RSMEXTRA:
         if (new_json_obj['feature_subset_file'] == default_feature_subset_file
-            and not new_json_obj['sign']):
+                and not new_json_obj['sign']):
             new_json_obj['sign'] = default_feature_sign
 
     # 6. Check for fields that must be specified together
@@ -387,7 +385,6 @@ def process_json_fields(json_obj):
     -------
     ValueError
     """
-
 
     list_fields = ['feature_prefix',
                    'general_sections',
@@ -529,6 +526,7 @@ def normalize_and_validate_feature_file(feature_json):
                          "in the feature file: {}".format(duplicate_features))
 
     return new_json_obj
+
 
 def read_json_file(json_file):
     """
@@ -764,6 +762,7 @@ def check_flag_column(config_obj):
 
     return new_filtering_dict
 
+
 def check_feature_subset_file(feature_specs, subset=None, sign=None):
     """
     Check that the file is in the correct format and contains all
@@ -796,9 +795,9 @@ def check_feature_subset_file(feature_specs, subset=None, sign=None):
             raise ValueError("The subset columns in feature file can only contain 0 or 1")
 
     if sign:
-        if not 'Sign_'+sign in feature_specs.columns:
+        if not 'Sign_{}'.format(sign) in feature_specs.columns:
                 raise ValueError("The feature_subset_file does not contain the requested "
-                                 "sign column {}".format('Sign_'+sign))
+                                 "sign column 'Sign_{}'".format(sign))
 
         if not feature_specs[subset].isin(['-', '+']).all():
             raise ValueError("The sign columns in feature file can only contain - or +")
@@ -879,7 +878,7 @@ def load_experiment_data(main_config_file, output_dir):
     exclude_listwise = False
     min_items = config_obj['min_items_per_candidate']
     if min_items:
-        exclude_listwise=True
+        exclude_listwise = True
 
     # get the name of the model that we want to train and
     # check that it's valid
@@ -920,7 +919,7 @@ def load_experiment_data(main_config_file, output_dir):
     if custom_report_section_paths:
         logger.info('Locating custom report sections')
         custom_report_sections = locate_custom_sections(custom_report_section_paths,
-                                                         configpath)
+                                                        configpath)
     else:
         custom_report_sections = []
 
@@ -958,7 +957,6 @@ def load_experiment_data(main_config_file, output_dir):
     else:
         feature_subset_specs = None
 
-
     # Do we need to automatically find the best transformations/change sign?
     select_transformations = config_obj['select_transformations']
     feature_sign = config_obj['sign']
@@ -987,12 +985,12 @@ def load_experiment_data(main_config_file, output_dir):
[…]
