### 2016-07-26 cac6063
Pep8 fixes
