+++ rsmtool/rsmeval.py
@@ -134,7 +134,7 @@ def run_evaluation(config_file, output_dir):
     exclude_listwise = False
     min_items_per_candidate = config_obj['min_items_per_candidate']
     if min_items_per_candidate:
-        exclude_listwise=True
+        exclude_listwise = True
 
     general_report_sections = config_obj['general_sections']
 
@@ -189,11 +189,11 @@ def run_evaluation(config_file, output_dir):
     #                  'scale', 'scale_trim' and 'scale_trim_round'.
 
     # we need to scale if and only if a CSV file is specified
-    do_scaling = (scale_with != None and scale_with != 'asis')
+    do_scaling = (scale_with is not None and scale_with != 'asis')
 
     # use scaled predictions for the analyses unless
     # we were told not to
-    use_scaled_predictions = (scale_with != None)
+    use_scaled_predictions = (scale_with is not None)
 
     # log an appropriate message
     if scale_with is None:
@@ -282,7 +282,6 @@ def run_evaluation(config_file, output_dir):
                                              'raw',
                                              'spkitemid',
                                              exclude_zeros=False)
-
 
     del df_filtered
     df_filtered_pred = newdf
@@ -554,4 +553,3 @@ def main():
 
 if __name__ == '__main__':
     main()
-
