### 2016-07-26 3985e80
Pep8 fixes
