### 2016-07-26 9dc28b1
[ci skip] add more tutorial.
