### 2016-07-26 cb5ff8c
minor typos and a note
