### 2016-07-26 a4affb7
[ci skip] Add rsmtool stdout.
