### 2016-07-26 8f8027b
[ci skip] more updates to tutorial.
