### 2016-07-26 7427cff
minor edits
