### 2016-07-26 25ece70
[ci skip] Update train/test files
