+++ doc/usage_rsmeval.rst
@@ -58,3 +58,5 @@ figure
 This folder contains all of the figures generated as part of the various analyses performed, saved as ``.svg`` files.
 
 .. include:: intermediate_files_rsmeval.rst
+
+.. include:: tutorial_rsmeval.rst
