### 2016-07-26 7f20493
tutorial for rsmeval
