### 2016-07-26 2745da4
Tutorials for rsmeval and rsmpredict
