### 2016-07-26 777635c
rsmcompare documentation
