+++ doc/tutorial_rsmcompare.rst
@@ -8,82 +8,78 @@ To use the ``rsmcompare`` you need to :ref:`install RSMTool <install>`.
 Workflow
 ~~~~~~~~
 
-RSMEval is designed for evaluating existing scores. Once you have the scores computed for all responses in your data, using RSMEval is fairly straightforward:
+``rsmcompare`` is designed to compare two existing ``rsmtool`` experiments. To use ``rsmcompare`` you need:
 
-1. Create a ``.csv`` file containing the system scores and the reference scores. 
-2. Create an :ref:`experiment configuration file <config_file_rsmeval>` describing the evaluation experiment you would like to run.
-3.  Run that configuration file with :ref:`rsmeval <usage_rsmeval>` and generate the experiment HTML report as well as the :ref:`intermediate CSV files <intermediate_files_rsmeval>`.
-4. Examine HTML report to check various aspects of model performance.
 
-Note that the above workflow does not use the customization features of RSMEval, e.g., :ref:`choosing which sections to include in the report <general_sections_rsmeval>` or :ref:`adding custom analyses sections <custom_notebooks>` etc. However, we will stick with this workflow for our tutorial since it is likely to be the most common use case.
+1. Run two experiments using :ref:`rsmtool <usage_rsmtool>` or :ref:`rsmeval <usage_rsmeval>`. 
+2. Create an :ref:`experiment configuration file <config_file_rsmcompare>` describing the comparison experiment you would like to run.
+3.  Run that configuration file with :ref:`rsmcompare <usage_rsmcompare>` and generate the comparison experiment HTML report.
+4. Examine HTML report to compare the two models.
+
+Note that the above workflow does not use the customization features of ``rsmcompare``, e.g., :ref:`choosing which sections to include in the report <general_sections_rsmcompare>` or :ref:`adding custom analyses sections <custom_notebooks>` etc. However, we will stick with this workflow for our tutorial since it is likely to be the most common use case.
 
 ASAP Example
 ~~~~~~~~~~~~
 
 We are going to use the same example from 2012 Kaggle competition on automated essay scoring that we used for :ref:`rsmtool tutorial <tutorial>`. 
 
-Generate scores 
-~~~~~~~~~~~~~~~
-RSMEval is designed for researchers who have developed their own system for generating scores and would like to produce an evaluation report for thse scores.  For this tutorial we will use the scores we generated for the evaluation set using :ref:`rsmtool tutorial <tutorial>`. 
+Run ``rsmtool`` experiments
+~~~~~~~~~~~~~~~~~~~~~~~~~~~
+``rsmcompare`` compares the results of the two existing ``rsmtool`` or ``rsmeval`` experiments. For this tutorial we will compare the outputs generated during the experiment in :ref:`rsmtool tutorial <tutorial>` to itself. If you have not already completed that tutorial first, please do so now. We will be using the output generated during the tutorial, so you will also need to complete the tutorial if you have already deleted the output files. 
 
 
 Create a configuration file
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~
-The next step is to create an :ref:`RSMEval experiment configuration file <config_file_rsmeval>` in ``.json`` format.
+The next step is to create an :ref:`rsmcompare experiment configuration file <config_file_rsmcompare>` in ``.json`` format.
 
-.. _asap_config_rsemeval:
+.. _asap_config_rsemcompare:
 
-.. literalinclude:: ../examples/rsmeval/config_rsmeval.json
+.. literalinclude:: ../examples/rsmeval/config_rsmcompare.json
     :language: javascript
     :linenos:
 
 Let's take a look at the options in our configuration file.
 
-- **Line 2**: We define an experiment ID
[…]
