### 2016-07-26 82c06dc
updated config.py
