### 2016-07-26 55bedff
[ci skip] Fixed typos and other minor changes.
- Also move tutorial to top for all tools.
