+++ doc/tutorial_rsmeval.rst
@@ -3,33 +3,31 @@
 Tutorial
 """"""""
 
-To use the rsmeval you need to :ref:`install RSMTool <install>`.
+For this tutorial, you first need to :ref:`install RSMTool <install>`.
 
 Workflow
 ~~~~~~~~
 
-RSMEval is designed for evaluating existing scores. Once you have the scores computed for all responses in your data, using RSMEval is fairly straightforward:
+``rsmeval`` is designed for evaluating existing machine scores. Once you have the scores computed for all the responses in your data, the next steps are fairly straightforward:
 
-1. Create a ``.csv`` file containing the system scores and the reference scores. 
+1. Create a ``.csv`` file containing the computed system scores and the human scores you want to compare against.
 2. Create an :ref:`experiment configuration file <config_file_rsmeval>` describing the evaluation experiment you would like to run.
 3.  Run that configuration file with :ref:`rsmeval <usage_rsmeval>` and generate the experiment HTML report as well as the :ref:`intermediate CSV files <intermediate_files_rsmeval>`.
-4. Examine HTML report to check various aspects of model performance.
+4. Examine the HTML report to check various aspects of model performance.
 
-Note that the above workflow does not use the customization features of RSMEval, e.g., :ref:`choosing which sections to include in the report <general_sections_rsmeval>` or :ref:`adding custom analyses sections <custom_notebooks>` etc. However, we will stick with this workflow for our tutorial since it is likely to be the most common use case.
+Note that the above workflow does not use any customization features , e.g., :ref:`choosing which sections to include in the report <general_sections_rsmeval>` or :ref:`adding custom analyses sections <custom_notebooks>` etc. However, we will stick with this workflow for our tutorial since it is likely to be the most common use case.
 
 ASAP Example
 ~~~~~~~~~~~~
+We are going to use the same example from 2012 Kaggle competition on automated essay scoring that we used for the :ref:`rsmtool tutorial <tutorial>`.
 
-We are going to use the same example from 2012 Kaggle competition on automated essay scoring that we used for :ref:`rsmtool tutorial <tutorial>`. 
-
-Generate scores 
+Generate scores
 ~~~~~~~~~~~~~~~
-RSMEval is designed for researchers who have developed their own system for generating scores and would like to produce an evaluation report for thse scores.  For this tutorial we will use the scores we generated for the evaluation set using :ref:`rsmtool tutorial <tutorial>`. 
-
+``rsmeval`` is designed for researchers who have developed their own scoring engine for generating scores and would like to produce an evaluation report for those scores.  For this tutorial, we will use the scores we generated for the ASAP2 evaluation set using :ref:`rsmtool tutorial <tutorial>`.
 
 Create a configuration file
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~
-The next step is to create an :ref:`RSMEval experiment configuration file <config_file_rsmeval>` in ``.json`` format.
+The next step is to create an :ref:`experiment configuration file <config_file_rsmeval>` in ``.json`` format.
 
 .. _asap_config_rsemeval:
 
@@ -39,32 +37,32 @@ The next step is to create an :ref:`RSMEval experiment configuration file <confi
 
 Let's take a look at the options in our configuration file.
 
-- **Line 2**: We define an experiment ID
+- **Line 2**: We define an experiment ID.
 - **Line 3**: We also provide a description which will be included in the experiment report.
-- **Line 4**: We list the path to the ``.csv`` with predicted scores.
-- **Line 5**: This field indicates that the system scores in our ``.csv`` file are located the columns named ``system``.
-- **Line 6**: This field indicates that the human (reference) scores in our ``.csv`` file are located the columns named ``human``
[…]
