### 2016-07-26 b6bdc52
[ci skip] Add note about conda environment to tutorials.
