### 2016-07-26 fdbf6b4
[ci skip] change path in log.
