### 2016-07-26 9bdec12
[ci skip] Rephrase conda bit
