### 2016-07-26 3c5c9c2
[ci skip] Prune README and remove .md files.
