+++ README.md
@@ -11,7 +11,7 @@ RSMTool is a python package which automates and combines in a single pipeline mu
 
 We expect the primary users of RSMTool to be researchers working on developing new automated scoring engines or on improving existing ones. Note that RSMTool is not a scoring engine by itself but rather a tool for building and evaluating machine learning models that may be used in such engines. 
 
-For details on usage and installation, please see the official [documentation](http://rsmtool.readthedocs.io). 
+For installation and usage, please see the official [documentation](http://rsmtool.readthedocs.io). 
 
 ## Requirements
 
@@ -32,91 +32,8 @@ For details on usage and installation, please see the official [documentation](h
 
 Contributions to RSMTool are very welcome. Please refer to the [documentation](http://rsmtool.readthedocs.io/en/latest/contributing.html) for how to get started on developing new features or functionality for RSMTool.
 
-
 ## Changelog
 See [GitHub Releases](https://github.com/EducationalTestingService/rsmtool/releases).
 
-**Use case 2**: A researcher *has* a proprietary automated scoring system (using a machine learning model that is not available in RSMTool) for grading short responses that extracts the features and computes the score. He wants to evaluate the system performance using analyses recommended by the educational measurement community that are not always available in standard machine learning packages. In addition, he also wants to conduct additional analyses to evaluate system fairness and compare whether his system agrees with the human grader as well as a second human grader would.
 
-He *uses* RSMTool to set up a customized evaluation report using a combination of existing and custom sections and then runs RSMEval to quickly produce a new report for each version of his system.
-
-
-RSMTool provides the following main scripts. The documentation for each of these tools is provided separately (see below). 
-
-* `rsmtool` - the tool for training and evaluating scoring models as described in Use case 1
-
-* `rsmeval` - the tool for evaluating predictions obtained from other systems as described in use case 2. 
-
-* `rsmpredict` - the tool for generating new predictions based on an existing models. For researchers who are already using `rsmtool` to build scoring model, this tool allows generating predictions for new data using an existing model. 
-
-* `rsmcompare` -  for comparing two `rsmtool` runs. This tool compares two models trained using `rsmtool`. It is most commonly used to compare model performance after adding new features.
-
-
-## Example
-
-You can try out RSMTool as follows:
-
-1. Go to the `example/rsmtool` folder. This folder contains the training and test set features for a simple scoring system built to automatically score the responses from the [2012 Kaggle Automated Student Assessment Prize competition](https://www.kaggle.com/c/asap-aes). 
-2. Make sure to activate the conda environment where you installed rsmtool (e.g., `source activate rsmtool`)
-3. We first try the whole pipeline by running RSMTool: `rsmtool config.json`
-4. Since no output directory was specfied, `rsmtool` will create the three output folders in the current directory: `figure`, `output`, and `report`. You can examine the HTML report `report/ASAP2_report.html`. It should look like [this](https://s3.amazonaws.com/sample-rsmtool-report/ASAP2_report.html).
-5. Now we will use `rsmpredict` to re-generate the scores for the test set without re-training the model. We will store these new predictions in `predictions.csv`: go to `../rsmpredict` and run `rsmpredict config_rsmpredict.json predictions.csv`. The tool will create a new file called `predictions.csv` which contains the predictions. 
-6. We will use `rsmeval` to evaluate these new predictions: go to `../rsmeval` and run `rsmeval config_rsmeval.json` .`rsmeval` will create the three output folders in the current directory: `figure`, `output`, and `report`. 
[…]
