+++ doc/rsmtool.md
@@ -1,49 +1,0 @@
-# rsmtool
-
-`rsmtool` preprocesses the training data (outlier truncation, normalization), builds a scoring models, and generates a report with the description of the model and in-depth evaluation of its performance on the evaluation data. 
-
-Most common use cases:
-
-- Train and evaluate a new scoring model for new data
-
-- Re-train an existing model on new data
-
-- Evaluate the model performance after adding a new feature
-
-- Generate descriptive statistics for feature distributions and correlations
-
-`rsmtool` contains a series of in-built models and also supports all regressors implemented in [SKLL](http://skll.readthedocs.org/en/latest/run_experiment.html#learners)(see [available models](available_models.md) for the full list).
-
-
-## Input
-
-`rsmtool` requires the following input:
-
-- The data files for training and evaluation sets. The files must be in .csv format. Each row should correspond to a single response and contain numeric feature values extracted for this response. In addition there should be a column with a unique id and a column with human score for each response. The column names are defined in the configuration file. 
-
-- Configuration file in .json format with the description of the data and the model. See [config_file.md](config_file.md) for further information. 
-
-In addition, if you want to use a specific set of features, you need to provide a second `.json` file specifying the list of features. See [feature_file.md](feature_file.md) for further information. 
-
-
-## Output
-
-`rsmtool` produces three folders with the following content: 
-
-- `figure/`: figures generated by the analyses
-
-- `output/`:  model outputs in .csv format (see [output_csv.md](output_csv.md))
-
-- `report/`:  model report as .html file and .ipnb python notebook. 
-
-
-## Usage
-
-`rsmtool config_file [output_directory]`
-
-- `config_file` - the path to the configuration file in .json file with the description of the data and model. 
-
-- `output-directory` - the directory where `rsmtool` will save its output. 
-Default: current folder
-
-- `--force` - by default, `rsmtool` will raise an Exception if you specify an output directory that already contains a directory called `output`. This is to ensure that an older experiment does not interfere with the current one. However, if you specify this flag, `rsmtool` assumes that you are an advanced user who knows what they are doing and only outputs a warning. 

[…]
