### 2016-07-26 e16639f
[ci skip] Remove extraneous info from environment.yml
