+++ environment.yml
@@ -1,4 +1,3 @@
-name: rsmtool
 channels:
   - defaults
   - desilinguist
@@ -79,5 +78,4 @@ dependencies:
 - xz=5.2.2=0
 - zeromq=4.1.3=0
 - zlib=1.2.8=3
-prefix: /Users/nmadnani/anaconda/envs/rsmtool
 
