### 2016-07-26 bc6df72
Minor edits
