### 2016-07-26 ed6ed7f
[ci skip] Switch to using pip for RTD.
