+++ environment.yml
@@ -1,81 +1,0 @@
-channels:
-  - defaults
-  - desilinguist
-dependencies:
-- alabaster=0.7.8=py34_0
-- appnope=0.1.0=py34_0
-- babel=2.3.3=py34_0
-- backports_abc=0.4=py34_0
-- beautiful-soup=4.3.2=py34_0
-- coverage=4.1=py34_0
-- cycler=0.10.0=py34_0
-- decorator=4.0.10=py34_0
-- prettytable=0.7.2=py34_2
-- ruamel.yaml=0.11.11=py34_0
-- skll=1.2.1=py34_0
-- docutils=0.12=py34_2
-- entrypoints=0.2.2=py34_0
-- freetype=2.5.5=1
-- imagesize=0.7.1=py34_0
-- ipykernel=4.3.1=py34_0
-- ipython=4.1.2=py34_2
-- ipython_genutils=0.1.0=py34_0
-- ipywidgets=4.1.1=py34_0
-- jinja2=2.8=py34_1
-- joblib=0.9.4=py34_0
-- jsonschema=2.5.1=py34_0
-- jupyter=1.0.0=py34_3
-- jupyter_client=4.3.0=py34_0
-- jupyter_console=5.0.0=py34_0
-- jupyter_core=4.1.0=py34_0
-- libpng=1.6.22=0
-- markupsafe=0.23=py34_2
-- matplotlib=1.5.1=np110py34_0
-- mistune=0.7.2=py34_1
-- nbconvert=4.2.0=py34_0
-- nbformat=4.0.1=py34_0
-- nomkl=1.0=0
-- nose=1.3.7=py34_1
-- notebook=4.1.0=py34_2
-- numpy=1.10.4=py34_nomkl_2
-- openssl=1.0.2h=1
-- pandas=0.18.0=np110py34_0
-- path.py=8.2.1=py34_0
-- patsy=0.4.1=py34_0
-- pexpect=4.0.1=py34_0
-- pickleshare=0.7.2=py34_0
-- pip=8.1.2=py34_0
-- prompt_toolkit=1.0.3=py34_0
-- ptyprocess=0.5.1=py34_0
-- pygments=2.1.3=py34_0
-- pyparsing=2.1.4=py34_0
-- pyqt=4.11.4=py34_4
-- python=3.4.5=0
-- python-dateutil=2.5.3=py34_0
-- pytz=2016.6.1=py34_0
-- pyzmq=15.3.0=py34_0
-- qt=4.8.7=4
-- qtconsole=4.2.1=py34_0
-- readline=6.2=2
-- scikit-learn=0.17.1=np110py34_nomkl_2
-- scipy=0.17.0=np110py34_nomkl_4
-- seaborn=0.7.0=py34_0
-- setuptools=23.0.0=py34_0
-- simplegeneric=0.8.1=py34_1
-- sip=4.18=py34_0
-- six=1.10.0=py34_0
-- snowballstemmer=1.2.1=py34_0
-- sphinx=1.4.1=py34_0
-- sphinx_rtd_theme=0.1.9=py34_0
-- sqlite=3.13.0=0
-- statsmodels=0.6.1=np110py34_1
-- terminado=0.6=py34_0
-- tk=8.5.18=0
-- tornado=4.4.1=py34_0
-- traitlets=4.2.2=py34_0
-- wcwidth=0.1.7=py34_0
-- wheel=0.29.0=py34_0
-- xz=5.2.2=0
-- zeromq=4.1.3=0
-- zlib=1.2.8=3
-
