### 2016-07-27 2c32840
Remove RTD pip file since we are using conda.
