+++ requirements_rtd.txt
@@ -1,17 +1,0 @@
-ipython==4.1.2
-jupyter==1.0.0
-joblib==0.9.4
-matplotlib==1.5.1
-nose==1.3.7
-notebook==4.1.0
-numpy==1.10.4
-pandas==0.18.0
-scikit-learn==0.17.1
-scipy==0.17.0
-seaborn==0.7.0
-skll==1.2.1
-statsmodels==0.6.1
-sphinx
-sphinx_rtd_theme
-coverage
-setuptools
