### 2016-07-27 bc8798b
Add docs badge to README.
