### 2016-07-27 2a59e8c
[ci skip] Fix some broken links.
