### 2016-07-27 #33 @desilinguist
Hi @arfon and @jkahn.

We have just released [v5.1.0](https://github.com/EducationalTestingService/rsmtool/releases/tag/v5.1.0) of RSMTool that addresses the suggestions made in the review. Specifically:

1. The documentation has been completely overhauled and is now hosted on [readthedocs](http://rsmtool.readthedocs.io). It now includes a very clear statement of need, several tutorials, as well as browsable API documentation.

2. (Almost all) expected warnings are now suppressed when running `nosetests`. One particular warnings remains which actually does indicate the use of about-to-be-deprecated code in a related package for which I have filed an [issue](https://github.com/EducationalTestingService/skll/issues/325). This should be fixed in the next release.

3. Several stylistic issues fixed using [pep8](https://github.com/EducationalTestingService/rsmtool/pull/40) and [pyflakes](https://github.com/EducationalTestingService/rsmtool/pull/39) except for a couple that are documented in the PRs.

4. Code coverage is now automatically computed via [coveralls.io](http://coveralls.io).

5. I have started working on `pip` compatibility and I think I have a way to get it working in the next release when we update one of the packages which has since become more wheel-friendly.
