### 2016-07-27 f1073aa
[ci skip] fix typos in readme.
