### 2016-07-27 2eb62cb
Corrected typo in paper.md
