### 2016-07-29 #33 @arfon
> I give this a 👍 and defer to the editors as to when/if/how you should mint a new DOI.

Excellent, thanks @jkahn.

@desilinguist - is there an associated DOI for the `v5.1.0` release. If so, please add the DOI as a comment here.
