+++ README.md
@@ -1,6 +1,6 @@
 [![Circle CI](https://circleci.com/gh/EducationalTestingService/rsmtool/tree/master.svg?style=shield)](https://circleci.com/gh/EducationalTestingService/rsmtool/tree/master)
 [![Coverage Status](https://coveralls.io/repos/github/EducationalTestingService/rsmtool/badge.svg?branch=feature%2Fadd-test-coverage)](https://coveralls.io/github/EducationalTestingService/rsmtool?branch=feature%2Fadd-test-coverage)
-[![DOI](https://zenodo.org/badge/22127/EducationalTestingService/rsmtool.svg)](https://zenodo.org/badge/latestdoi/22127/EducationalTestingService/rsmtool)
+[![DOI](https://zenodo.org/badge/doi/10.5281/zenodo.58851.svg)](http://dx.doi.org/10.5281/zenodo.58851)
 [![GitHub license](https://img.shields.io/badge/license-Apache%202-blue.svg)](https://raw.githubusercontent.com/EducationalTestingService/rsmtool/master/LICENSE)
 [![Documentation Status](https://readthedocs.org/projects/rsmtool/badge/?version=latest)](http://rsmtool.readthedocs.io/en/latest/?badge=latest)
 
