### 2016-07-29 #33 @desilinguist
Hi @arfon, here is the DOI for the `v5.1.0` release

[![DOI](https://zenodo.org/badge/doi/10.5281/zenodo.58851.svg)](http://dx.doi.org/10.5281/zenodo.58851)

Please let me know if that's okay or if I need to do something else.
