### 2016-07-29 #33 @arfon
Perfect. Thanks!

@desilinguist - your paper is now accepted into JOSS and you DOI is http://dx.doi.org/10.21105/joss.00033

Thanks for the great review @jkahn
